% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{align_titles}
\alias{align_titles}
\title{Horizontally align ggplot2 title and subtitle}
\usage{
align_titles(align = 0)
}
\arguments{
\item{align}{Horizontal alignment. \code{0} left-aligns, \code{1} right-aligns, \code{0.5} centers.}
}
\value{
A ggplot2 theme object.
}
\description{
Horizontally align ggplot2 title and subtitle
}
\examples{

# library(ggplot2)
# ggplot(mpg, aes(manufacturer, cty)) + geom_boxplot() +
#     labs(title = "This is a title", subtitle = "This is a subtitle") +
#     align_titles(align = 0.5)

#> A plot with the title and subtitle centered.

}
