% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_tools.R
\name{assign_groups}
\alias{assign_groups}
\title{Assign elements in a vector to groups}
\usage{
assign_groups(vec, g, balance = TRUE, dedupe = TRUE)
}
\arguments{
\item{vec}{(Numeric or Character) A vector. No sorting is done to the vector
so if you want to group based on some kind of ordering, you need to do it
beforehand.}

\item{g}{(Integer) The maximum of of groups to split \code{vec} into (the maximum
group size may not be reached, e.g. grouping 4 elements into 6 groups).}

\item{balance}{(Logical) If \code{TRUE}, try to have equal numbers of observations
per group.}

\item{dedupe}{(Logical) If \code{TRUE}, duplicate values in \code{vec} will be ignored
when generating the groups, ensuring that identical values go into the
same group. This can unbalance the groups even if \code{balance = TRUE}.}
}
\value{
An integer vector of the same length as \code{vec}.
}
\description{
Assign elements in a vector to groups
}
\examples{
testvec <- c(4, 7, 8, 2, 2, 2, 5, 1, 6, 3)

# 10 values assigned to 4 balanced groups.
assign_groups(testvec, 4, balance = TRUE, dedupe = FALSE)
#> [1] 1 1 1 2 2 2 3 3 3 4

# 10 values assigned to 4 unbalanced groups.
assign_groups(testvec, 4, balance = FALSE, dedupe = FALSE)
#> [1] 1 1 1 1 2 2 3 3 4 4

# 8 values (plus 2 duplicates) assigned to 4 groups.
assign_groups(testvec, 4, balance = TRUE, dedupe = TRUE)
#> [1] 1 1 2 2 2 2 3 3 4 4

}
