% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exports.R
\docType{data}
\name{basic_colour_names}
\alias{basic_colour_names}
\alias{basic_color_names}
\title{The names of 197 common colours}
\description{
This is a character vector of common colour names, lowercased and sorted by word length
so that common colours like red and blue come earlier in the list.
}
\section{Author}{

\itemize{
\item Desi Quintans (@eco_desi)
}
}

\section{Source}{

\itemize{
\item \url{https://simple.wikipedia.org/wiki/Web_color}
\item \url{https://simple.wikipedia.org/wiki/List_of_colors}
}
}

\examples{
# data(basic_color_names)

# print(basic_color_names)

#>   [1] "red"                    "tan"                    "aqua"
#>   [4] "blue"                   "cyan"                   "erin"
#>   [7] "gold"                   "gray"                   "grey"
#>  [10] "jade"                   "lime"                   "navy"
#>  [13] "pear"                   "peru"                   "pink"
#>  [16] "plum"                   "puce"                   "rose"
#>  [19] "ruby"                   "snow"                   "teal"
#>  [22] "amber"                  "azure"                  ...

#> data(basic_colour_names)

}
