% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_tools.R
\name{breadcrumb}
\alias{breadcrumb}
\title{Return the stem that is common to a set of strings}
\usage{
breadcrumb(..., na.rm = FALSE)
}
\arguments{
\item{...}{(Vectors) Vectors that will be coerced into Character and
appended together.}

\item{na.rm}{(Logical) Should \code{NA} be removed from the input vectors?}
}
\value{
A string.
}
\description{
Return the stem that is common to a set of strings
}
\section{Authors}{

\itemize{
\item Desi Quintans (\url{http://www.desiquintans.com})
}
}

\examples{
breadcrumb("volunteer", "volunteering", "voluntourism", "volunteers", na.rm = TRUE)
#> [1] "volunt"

# The function does not return substrings:

breadcrumb("able", "wobble")
#> character(0)

}
