% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_funs.R
\name{build_palette}
\alias{build_palette}
\title{Build a palette of colours from a list of hex codes}
\usage{
build_palette(col_list, n = NULL, random = FALSE, alpha = NULL)
}
\arguments{
\item{col_list}{(Character) A vector of colours in RGB Hex format without transparency.}

\item{n}{(Numeric or \code{NULL}) The number of colours to return. If \code{NULL}, return all of
the colours for \code{col_list}.}

\item{random}{(Logical) If \code{TRUE}, colours will be randomly sampled without replacement.
If \code{FALSE}, they will be drawn from \code{col_list} in order.}

\item{alpha}{A single decimal value that modifies the opacity of the colours.
\code{alpha = 0.65} makes all of the colours 65 percent opaque (35 percent transparent).
If \code{alpha = NULL}, no alpha channel is added and the colours are 100 percent opaque.}
}
\value{
A character vector of hex colours. If \code{alpha = NULL}, the colours will be in
RGB Hex format (e.g. #FFFF00). If \code{alpha} is not \code{NULL}, the colours will be in
RGBA Hex format (e.g. #FFFF00CB).
}
\description{
Build a palette of colours from a list of hex codes
}
\examples{
\dontrun{
build_palette(c("#000000", "#FFFF00", "#1CE6FF", "#FF34FF", "#FF4A46"), alpha = 0.5)
#> [1] "#00000080" "#FFFF0080" "#1CE6FF80" "#FF34FF80" "#FF4A4680"
}

}
