% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_tools.R
\name{cat_wrap}
\alias{cat_wrap}
\title{Print to console, wrapping the text to a specific line width}
\usage{
cat_wrap(text, width = 80, element_sep = "\\n")
}
\arguments{
\item{text}{(Character) A vector of text.}

\item{width}{(Integer) The character length of each line.}

\item{element_sep}{(Character) A string to insert between each element of \code{text}.}
}
\value{
Print to console.
}
\description{
Wrapping console output is essential in Rmarkdown documents because long character
vectors do not wrap when printed inside code blocks.
}
\section{Authors}{

\itemize{
\item Desi Quintans (\url{http://www.desiquintans.com})
}
}

\examples{
vec <- c("This is a very long chunk of text.",
         "This is also another quite long chunk of text.")

cat_wrap(vec, width = 25)

#> This is a very long
#> chunk of text.
#>
#> This is also another
#> quite long chunk of
#> text.

}
