% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_tools.R
\name{coinflip}
\alias{coinflip}
\title{Randomly return TRUE or FALSE}
\usage{
coinflip(n = 1)
}
\arguments{
\item{n}{(Integer) The number of times to 'flip the coin'}
}
\value{
Returns \code{TRUE} or \code{FALSE} randomly.
}
\description{
Randomly return TRUE or FALSE
}
\section{Authors}{

\itemize{
\item Desi Quintans (\url{http://www.desiquintans.com})
}
}

\examples{
coinflip()
#> [1] FALSE

coinflip(5)
#> [1] TRUE TRUE FALSE TRUE FALSE

}
