% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_tools.R
\name{collapse_df}
\alias{collapse_df}
\title{Collapse a dataframe into a vector}
\usage{
collapse_df(df, from = 1, to = NULL, cols = NULL)
}
\arguments{
\item{df}{(Dataframe) A dataframe.}

\item{from, to}{(Numeric or \code{NULL}) The start and end of a continuous range of columns
that will be considered for the empty/not-empty decision. For example, columns that
are always filled should be omitted (see examples). If \code{to} is \code{NULL}, it defaults
to the last column in \code{df} so that \code{from = 2, to = NULL} is the same as
\code{2:length(df)}.}

\item{cols}{(Numeric or \code{NULL}) A numeric vector of the columns to consider. This
allows you to select non-contiguous columns. If the \code{cols} argument is being used
(not-\code{NULL}), \code{from} and \code{to} will be ignored.}
}
\value{
A vector containing the cell contents from the selected columns of \code{df}.
If all of the cells are numeric, the vector is Numeric. If any of the cells contain
strings, the vector is Character. The columns are concatenated in order.
}
\description{
Useful for taking every number in a table and plotting it in a histogram, for example.
}
\section{Authors}{

\itemize{
\item Desi Quintans (\url{http://www.desiquintans.com})
}
}

\examples{
collapse_df(iris, cols = 1:4)

#> [1] 5.1 4.9 4.7 4.6 5.0 5.4 4.6 5.0 4.4 4.9 5.4 4.8 ...

}
