% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_tools.R
\name{common_stem}
\alias{common_stem}
\title{Return the stem that is common to a set of strings}
\usage{
common_stem(..., side = "left", na.rm = FALSE)
}
\arguments{
\item{...}{(Vectors) Vectors that will be coerced into Character and
appended together.}

\item{side}{(Character) Search from the \code{left} or \code{right} side of the strings.
Also accepts \code{l} or \code{r} as shorthand.}

\item{na.rm}{(Logical) Should \code{NA} be removed from the input vectors?}
}
\value{
A string. If there is no common stem among all the words, an empty string
of length 1 ("") will be returned.
}
\description{
Return the stem that is common to a set of strings
}
\section{Authors}{

\itemize{
\item Desi Quintans (\url{http://www.desiquintans.com})
}
}

\examples{
vec <- c("exciting", "exceeding", "excepting")

common_stem(vec)
#> [1] "exc"

common_stem(vec, side = "r")
#> [1] "ing"

# The function does not return substrings:
common_stem("tableaux", "wobbles")
#> ""

}
