% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_tools.R
\name{concat_nums}
\alias{concat_nums}
\title{Concatenate numbers together}
\usage{
concat_nums(...)
}
\arguments{
\item{...}{Numbers, or strings that can be coerced to numbers.}
}
\value{
A double.
}
\description{
Paste a set of numbers side-by-side. Useful for manually building ID numbers from
fields. Missing values are replaced with 0.
}
\section{Authors}{

\itemize{
\item Desi Quintans (\url{http://www.desiquintans.com})
}
}

\examples{
concat_nums(12, "76", NA, 1.5)

#> [1] 127601.5

}
