% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime_tools.R
\name{consecutive_month}
\alias{consecutive_month}
\title{Find the current month number relative to a starting date}
\usage{
consecutive_month(from, to)
}
\arguments{
\item{from}{(Character or Numeric) A date in the ISO-8601 form \code{"YYYY-MM-DD"}, or a
single number for the year which implies \code{YYYY-01-01}.}

\item{to}{(Character) A date in the ISO-8601 form \code{YYYY-MM-DD}.}
}
\value{
A numeric vector for the number of months elapsed.
}
\description{
Sometimes it's helpful to be able to break down elapsed time in more granular
units, like being able to say, "I am in the 15th month of this experiment."
}
\section{Authors}{

\itemize{
\item Desi Quintans (\url{http://www.desiquintans.com})
}
}

\examples{
# These are the same thing.
consecutive_month(2015, "2016-02-04")
consecutive_month("2015-01-01", "2016-02-04")
#> [1] 14

consecutive_month("2015-02-01", "2016-02-04")
#> [1] 13

}
