% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime_tools.R
\name{consecutive_week}
\alias{consecutive_week}
\title{Find the number of weeks between two dates}
\usage{
consecutive_week(from, to, tz = NULL, iso = FALSE)
}
\arguments{
\item{from}{(Character or Numeric) A date in the ISO-8601 form \code{"YYYY-MM-DD"}, or a
single number for the year which implies \code{YYYY-01-01}.}

\item{to}{(Character) A date in the ISO-8601 form \code{YYYY-MM-DD}.}

\item{tz}{(Character or \code{NULL}) A timezone name such as "Australia/Sydney".
Use \code{OlsonNames()} for a list of these.}

\item{iso}{(Logical) If \code{FALSE}, returns a naive estimate of weeks elapsed.
With R's \code{difftime()}, it's the same as dividing the days elapsed by 7.
Use \code{round()}, \code{ceiling()}, or \code{floor()} to simplify the result if desired.
\cr
If \code{TRUE}, returns the number of isoweeks (calendar weeks starting on
Mondays) between the two dates. This means that yesterday (Sunday) and
today (Monday) fall across two isoweeks even though they're only one
day apart.}
}
\value{
A numeric vector for the number of weeks elapsed.
}
\description{
Sometimes it's helpful to be able to break down elapsed time in more granular
units, like being able to say, "This experiment has been running for
15 weeks". It may also be helpful to say, "It has been 10 calendar weeks since
this happened". This function does both of those.
}
\section{Authors}{

\itemize{
\item Desi Quintans (\url{http://www.desiquintans.com})
}
}

\examples{
consecutive_week("2022-01-01", "2022-05-02", iso = FALSE)
#> [1] 17.28571

# Same dates, but they fall over multiple calendar weeks.
consecutive_week("2022-01-01", "2022-05-02", iso = TRUE)
#> [1] 4

# Note that consecutive days can spill over isoweeks.
consecutive_week("2022-05-01", "2022-05-02", iso = TRUE)
#> [1] 2


}
