% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_funs.R
\name{construct_cols}
\alias{construct_cols}
\title{Subset dataframe columns}
\usage{
construct_cols(df, from = 1, to = NULL, cols = NULL)
}
\arguments{
\item{df}{(Dataframe) A dataframe.}

\item{from, to}{(Numeric or \code{NULL}) The start and end of a continuous range of columns
that will be considered for the empty/not-empty decision. For example, columns that
are always filled should be omitted (see examples). If \code{to} is \code{NULL}, it defaults
to the last column in \code{df} so that \code{from = 2, to = NULL} is the same as
\code{2:length(df)}.}

\item{cols}{(Numeric or \code{NULL}) A numeric vector of the columns to consider. This
allows you to select non-contiguous columns. If the \code{cols} argument is being used
(not-\code{NULL}), \code{from} and \code{to} will be ignored.}
}
\value{
A numeric vector that can be used to reference columns in \code{df}.
}
\description{
This function mostly exists so that I don't have to keep typing \code{df[2:length(df)]} or
some other variant of it. It lets me subset columns based on contiguous or
non-contiguous ranges.
}
