% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_tools.R
\name{count_unique}
\alias{count_unique}
\title{Count how many times each unique element in a vector is repeated}
\usage{
count_unique(..., sort = FALSE)
}
\arguments{
\item{...}{(Vectors) Vectors that will be concatenated together.}

\item{sort}{(Logical) If \code{TRUE}, the results will be sorted by decreasing count.}
}
\value{
A dataframe with two columns: \code{unique} which lists the unique value, and
\code{count} which shows how many times that unique value appeared in \code{...}.
}
\description{
Count how many times each unique element in a vector is repeated
}
\examples{
count_unique(sample(letters, size = 10, replace = TRUE))

#>    unique  count
#> 1       e      1
#> 2       g      1
#> 3       i      2
#> 4       m      1
#> 5       n      1
#> 6       o      1
#> 7       p      2
#> 8       y      1

}
