% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_tools.R
\name{cumpct}
\alias{cumpct}
\title{Cumulative percentage}
\usage{
cumpct(nums)
}
\arguments{
\item{nums}{(Numeric) A vector of numbers that will be used in the order they
are provided.}
}
\value{
A Numeric vector.
}
\description{
Cumulative percentage
}
\examples{
cumpct(rep(1, 10))

## [1] 0.1 0.2 0.3 0.4 0.5 0.6 0.7 0.8 0.9 1.0

}
