% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_tools.R
\name{cut_int}
\alias{cut_int}
\title{Cuts for dividing an integer into \code{n} chunks}
\usage{
cut_int(int, chunks = NULL)
}
\arguments{
\item{int}{(Integer) The number to divide.}

\item{chunks}{(Integer) The number of chunks to divide it into.}
}
\value{
A tibble with four columns: \code{chunk}, which is the chunk's number, \code{left},
which is the start of each chunk, \code{right}, which is its end, and \code{size}, which
is how many items are in the chunk.
}
\description{
This function is intended for things like dividing large datasets into \code{n} chunks
with a more-or-less equal number of rows in each chunk. Useful for generating batch
files.
}
\section{Authors}{

\itemize{
\item Desi Quintans (\url{http://www.desiquintans.com})
}
}

\examples{
cut_int(19.4e6, 9)

# # A tibble: 9 × 4
#   chunk     left    right    size
#   <int>    <dbl>    <dbl>   <dbl>
#       1        1  2155557 2155557
#       2  2155558  4311112 2155555
#       3  4311113  6466668 2155556
#       4  6466669  8622223 2155555
#       5  8622224 10777779 2155556
#       6 10777780 12933334 2155555
#       7 12933335 15088890 2155556
#       8 15088891 17244445 2155555
#       9 17244446 19400000 2155555

}
