% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_tools.R
\name{degreedays}
\alias{degreedays}
\title{Calculate degree-days}
\usage{
degreedays(min, max, base, method = "modavg")
}
\arguments{
\item{min}{(Numeric) The lowest temperature in the day.}

\item{max}{(Numeric) The highest temperature in the day.}

\item{base}{(Numeric) The threshold for counting degree-days. This could be the
temperature when plants begin their spring growth or insects become noticeably
active, or some other biologically-meaningful baseline.}

\item{method}{(String, \code{avg} or \code{modavg}) The calculation method to use (see above).
Default is \code{modavg}.}
}
\value{
A numeric vector of the same length as min and max, with each entry containing
the degree-day value for that day.
}
\description{
"In a nutshell: heating degree days are a measure of how much (in degrees), and for
how long (in days), the air temperature was below a certain level." --- Martin
Bromley, \url{http://www.degreedays.net/introduction}
}
\details{
This function exposes two calculation methods that were taken from

Herms, Daniel A. 2013. “Using Degree-Days and Plant Phenology to Predict Pest
Activity.” IPM of Midwest Landscapes: Tactics and Tools for IPM. 2013.
http://cues.cfans.umn.edu/old/IPM-Tactics/IPM-tactics.html.

\strong{Average method (\code{avg})}

"If the maximum temperature for the day never rises above the base temperature, then
no development occurs, and zero degree-days accumulate (we don't calculate negative
degree-day values since the development of organisms does not reverse when it is
cold)." --- Daniel A. Herms

\strong{Modified Average (\code{modavg})}

"When the daily temperature fluctuates above and below the base temperature (as it
does frequently in the spring), the Average Method can underestimate the number of
degree-days actually experienced by a plant or insect. In this situation, the Modified
Average Method will calculate a higher number of degree-days, and thus can be more
accurate for predicting pest activity than the Average Method." --- Daniel A. Herms
}
\section{Authors}{

\itemize{
\item Desi Quintans (\url{http://www.desiquintans.com})
\item Daniel A. Herms \url{https://scholar.google.com/citations?user=yMYvlKsAAAAJ&hl=en}
}
}

\section{Source}{

Herms, Daniel A. 2013. “Using Degree-Days and Plant Phenology to Predict Pest
Activity.” IPM of Midwest Landscapes: Tactics and Tools for IPM. 2013.
http://cues.cfans.umn.edu/old/IPM-Tactics/IPM-tactics.html.
}

\examples{
degreedays(min = c(19, 20, 20, 21), max = c(25, 24, 23, 22), base = 22, method = "modavg")

#> [1] 1.5 1.0 0.5 0.0

}
