% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{dendro}
\alias{dendro}
\title{Make a dendrogram}
\usage{
dendro(df, clust_method = "ave", dist_method = "euclidean", labels = NULL,
  main = NULL, sub = NULL, cex = 0.8)
}
\arguments{
\item{df}{(Dataframe or Matrix) The data, arranged in wide format. Non-numeric data
will be coerced into \code{NA}s.}

\item{clust_method}{(Character) the agglomeration method to be used. This should be
(an unambiguous abbreviation of) one of "ward.D", "ward.D2", "single", "complete",
"average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC) or "centroid"
(= UPGMC).}

\item{dist_method}{(Character) the distance measure to be used. This must be one of
"euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski". Any
unambiguous substring can be given.}

\item{labels}{(Character) Text to use as the tip labels.}

\item{main}{(Character) Chart title.}

\item{sub}{(Character) Chart subtitle.}

\item{cex}{(Numeric) Size magnification. \code{0.8} is 80 percent of the base value.}
}
\value{
A dendrogram plot.
}
\description{
Make a dendrogram
}
\examples{
\dontrun{
df <- data.frame(stringsAsFactors = FALSE,
                 site = c("a", "b", "c", "d"),
                 spp1 = c(0L, 1L, 1L, 0L),
                 spp2 = c(1L, 1L, 1L, 0L),
                 spp3 = c(0L, 0L, 1L, 0L))

dendro(df[2:4], labels = df$site)
}

}
