% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desiderata.R
\docType{package}
\name{desiderata}
\alias{desiderata}
\title{desiderata - A pack of personal functions}
\section{Author}{

Desi Quintans (@eco_desi)
}

\section{URL}{

\url{https://github.com/DesiQuintans/taxizehelper/}
}

