% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_tools.R
\name{diff_cols}
\alias{diff_cols}
\title{Given two dataframes, which columns are present in one but not in the other?}
\usage{
diff_cols(l, r, side = "both")
}
\arguments{
\item{l}{(Dataframe) A dataframe whose column names to compare.}

\item{r}{(Dataframe) A dataframe whose column names to compare.}

\item{side}{(Character) \code{"both"} or \code{"b"} (default) finds columns that are missing from
both dataframes. \code{"left"} or \code{"l"} finds cols in \code{l} that are not in \code{r}.
\code{"right"} or \code{"r"} finds cols in \code{r} that are not in \code{l}.}
}
\value{
A Character vector with the names of missing columns.
}
\description{
Unlike \code{same_cols()}, the order of \code{l} and \code{r} \emph{does} matter for \code{diff_cols()}.
}
\section{Authors}{

\itemize{
\item Desi Quintans (\url{http://www.desiquintans.com})
}
}

\examples{
iris1 <- iris[, 1:3]
colnames(iris1)
## [1] "Sepal.Length" "Sepal.Width"  "Petal.Length"

iris2 <- iris[, 2:5]
colnames(iris2)
## [1]                "Sepal.Width"  "Petal.Length" "Petal.Width"  "Species"

diff_cols(iris1, iris2)
#> [1] "Sepal.Length" "Petal.Width"  "Species"

diff_cols(iris1, iris2, side = "l")
#> [1] "Sepal.Length"

diff_cols(iris1, iris2, side = "r")
#> [1] "Petal.Width"  "Species"

}
