% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_tools.R
\name{encode_signif}
\alias{encode_signif}
\title{Replace p-values with significance codes}
\usage{
encode_signif(p, codes = c("****", "***", "**", "*", "^", ""))
}
\arguments{
\item{p}{(Numeric) A vector of p-values.}

\item{codes}{(Character) A vector of 6 codes to use if \code{p <= threshold}. The
thresholds (in order) are: 0.0001, 0.001, 0.01, 0.05, 0.10, and 1.}
}
\value{
A Character vector.
}
\description{
Replace p-values with significance codes
}
\examples{
encode_signif(c(0, 0.001, 0.01, 0.05, 0.10))

## [1] "****" "***"  "**"   "*"    ""

}
