% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_tools.R
\name{ensure}
\alias{ensure}
\title{Quick assertion function for sanity checking your code}
\usage{
ensure(desc, expr)
}
\arguments{
\item{desc}{(Character) Human-readable description of the expression
being evaluated.}

\item{expr}{(Expression) The expression being tested. It should evaluate to
\code{TRUE} when the test passes, and \code{FALSE} when the test fails.}
}
\value{
An Error if the test fails, or a Message if the test passes.
}
\description{
This function can be used to do quick checks on your code, especially in
RMarkdown analysis documents. For example, making sure that the number of
rows in a dataframe hasn't changed after a left-join operation. For more
than a handful of tests, look for a fuller testing package like \code{assertr}.
}
\examples{
\dontrun{
letters
length(letters)

ensure("There should be 26 letters", length(letters) == 26)
}

}
