% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_funs.R
\name{find_dims}
\alias{find_dims}
\title{Find the nearest rectangle that will fit the number of entries in a vector}
\usage{
find_dims(vec)
}
\arguments{
\item{vec}{(Vector) A vector.}
}
\value{
A named numeric vector with 2 entries: \code{x} and \code{y} describing the number of
columns and rows respectively.
}
\description{
Find the nearest rectangle that will fit the number of entries in a vector
}
\examples{
find_dims(1:5)

#> x y
#> 2 3

}
