% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_tools.R
\name{fold}
\alias{fold}
\title{Friendly printing of a multi-element vector}
\usage{
fold(vec, n = 2)
}
\arguments{
\item{vec}{(Vector) The vector you want to print.}

\item{n}{(Integer) The maximum number of vector elements to show.}
}
\value{
A character vector of length 1 containing the readout (or \code{NA}).
}
\description{
Produces a readout like \code{"January, February, and 10 others"}.
}
\section{Authors}{

\itemize{
\item Desi Quintans (\url{http://www.desiquintans.com})
}
}

\examples{
month.name
#> [1] "January"   "February"  "March"     "April"     "May"       "June"
#> [7] "July"      "August"    "September" "October"   "November"  "December"

fold(month.name)
#> January, February, and 10 others

fold(1:20, n = 5)
#> 1, 2, 3, 4, 5, and 15 others

fold(character(0), n = 1)
#> [1] NA

fold(1:10, n = 0)
#> [1] NA

}
