% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_tools.R
\name{geomean}
\alias{geomean}
\title{Geometric mean of a vector}
\usage{
geomean(x, na.rm = TRUE)
}
\arguments{
\item{x}{(Numeric) A vector.}

\item{na.rm}{(Logical) If TRUE, NAs will be silently removed.}
}
\value{
A numeric vector of length 1.
}
\description{
The geometric mean is useful when your values are imbalanced by outliers. Daniel
McNichol wrote \href{https://towardsdatascience.com/on-average-youre-using-the-wrong-average-geometric-harmonic-means-in-data-analysis-2a703e21ea0}{a very approachablearticle}
about this topic.
}
\details{
This function is by Paul McMurdie and is "vectorized, zero- and NA-tolerant \link{...}
checks for any negative values, and returns a more informative and appropriate NaN
respecting that geometric mean is not defined for negative values (but is for zeros)."
}
\section{Authors}{

\itemize{
\item Paul McMurdie (\url{https://stackoverflow.com/users/935950/paul-mcmurdie})
}
}

\section{Source}{

\url{https://stackoverflow.com/a/25555105/5578429}
}

\examples{
vec <- c(1, 3, 9, 27, 81, 243, 729)

mean(vec)
#> [1] 156.1429

gm_mean(vec)
#> [1] 27

}
