% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_tools.R
\name{howmany}
\alias{howmany}
\title{Number of unique entries in a vector}
\usage{
howmany(...)
}
\arguments{
\item{...}{(Vector) Multiple vectors will be concatenated.}
}
\value{
The number of unique values in \code{...}.
}
\description{
Shorthand for \code{length(unique(x))}.
}
\examples{
my_vec <- sample(month.abb, 10, replace = TRUE)
#> [1] "Sep" "Nov" "Nov" "Jul" "Oct" "May" "Jun" "Nov" "Apr" "Jan"

howmany(my_vec)
#> 6

howmany(letters, LETTERS)
#> 52
}
