% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_tools.R
\name{howmany_df}
\alias{howmany_df}
\title{Count the number of unique values per column in a dataframe}
\usage{
howmany_df(df, ...)
}
\arguments{
\item{df}{(Dataframe) A dataframe, or something that can be coerced into a dataframe.}

\item{...}{(Names) Columns names or positions that are passed to \code{dplyr::select()}.}
}
\value{
A dataframe.
}
\description{
It's like \code{howmany()}, but for dataframes instead of vectors.
}
\section{Authors}{

\itemize{
\item Desi Quintans (\url{http://www.desiquintans.com})
}
}

\examples{
howmany_df(mtcars)
howmany_df(mtcars, -mpg)
howmany_df(mtcars, drat)

# Also tries to handle things that aren't dataframes, like tables.
howmany_df(Titanic)

}
