% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_tools.R
\name{IQR_outliers}
\alias{IQR_outliers}
\title{Mark outliers using the 1.5 * IQR method}
\usage{
IQR_outliers(x, multi = 1.5)
}
\arguments{
\item{x}{(Numeric) Vector to analyse.}

\item{multi}{(Numeric) Multiplication factor for the IQR. \code{1.5} by default.}
}
\value{
A Numeric vector that contains elements of \code{x} that are outliers.
}
\description{
Mark outliers using the 1.5 * IQR method
}
\section{Authors}{

\itemize{
\item 42- (https://stackoverflow.com/users/1855677/42)
}
}

\section{Source}{

\url{https://stackoverflow.com/a/49090426/5578429}
}

\examples{
IQR_outliers(c(-3000, 1:100, 1000))

## [1] -3000  1000

}
