% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exports.R
\docType{data}
\name{iris_labelled}
\alias{iris_labelled}
\title{The classic Iris dataset, with variable labels}
\format{
## `iris_labelled`
A data frame with 150 rows and 5 columns:
\describe{
  \item{Sepal.Length}{Length of sepals.}
  \item{Sepal.Width}{Width of sepals.}
  \item{Petal.Length}{Length of petals.}
  \item{Petal.Width}{Width of petals.}
  \item{Species}{Species of iris: setosa, versicolor, virginica}
  ...
}
}
\source{
R `datasets` package
}
\usage{
iris_labelled
}
\description{
The classic Iris dataset, with variable labels
}
\keyword{datasets}
