% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_tools.R
\name{is.flanked}
\alias{is.flanked}
\title{Is a vector element surrounded by certain values?}
\usage{
is.flanked(vec, items, edges_as_na = TRUE)
}
\arguments{
\item{vec}{(Vector) A vector.}

\item{items}{(Vector) A vector.}

\item{edges_as_na}{(Logical) If \code{TRUE} (default), the computation will add \code{NA} to the
start and end of \code{vec} \emph{temporarily}. If \code{FALSE}, it will reuse the first and last
values. See examples.}
}
\value{
A Logical vector of the same length as \code{vec}.
}
\description{
Checks if an element in a vector is flanked, i.e. if the elements before and after it
are \code{\%in\%} a vector of candidates.
}
\section{Authors}{

\itemize{
\item Desi Quintans (\url{http://www.desiquintans.com})
}
}

\examples{
test_vec <- c(1, NA, 2, NA, 3, 4, NA, 5)

is.flanked(test_vec, items = c(NA), edges_as_na = TRUE)

# The edges are regarded as flanked by `NA` because `edges_as_na == TRUE` adds `NA` to
# the start and end of `vec`.
#> [1]  TRUE FALSE  TRUE FALSE FALSE FALSE FALSE  TRUE


is.flanked(test_vec, items = c(NA), edges_as_na = FALSE)

#> [1] FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE

}
