% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_tools.R
\name{is.prime}
\alias{is.prime}
\title{Check if an integer is a prime number}
\usage{
is.prime(num)
}
\arguments{
\item{num}{(Numeric) An integer.}
}
\value{
\code{TRUE} if \code{num} is prime.
}
\description{
Check if an integer is a prime number
}
\section{Authors}{

\itemize{
\item flodel (\url{https://stackoverflow.com/users/1201032/flodel})
\item geotheory (\url{https://stackoverflow.com/users/1156245/geotheory})
\item Desi Quintans (<http://www.desiquintans.com)
}
}

\section{Source}{

\url{https://stackoverflow.com/a/19767707/5578429}
}

\examples{
is.prime(2)
#> [1] TRUE

is.prime(3)
#> [1] TRUE

is.prime(4)
#> [1] FALSE

is.prime(5)
#> [1] TRUE

}
