% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_tools.R
\name{loadRDS}
\alias{loadRDS}
\title{Load an RDS file and announce when it was created}
\usage{
loadRDS(...)
}
\arguments{
\item{...}{(Character) Arguments to send to \code{file.path()}. You can provide a complete
path as a single string, or incrementally build a path with many strings.}
}
\value{
An imported RDS file.
}
\description{
When you load a RDS file (serialised R objects e.g. a saved dataframe), this function
announces when the file was created so that you can see whether you are using an
up-to-date version. Often used in Rmarkdown documents.
}
\section{Authors}{

\itemize{
\item Desi Quintans (\url{http://www.desiquintans.com})
}
}

\examples{
# data <- loadRDS("path", "to", "data.rds")

#> Loading data.rds
#> It was compiled on 2018-05-16 11:36:05.

}
