% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_tools.R
\name{mark_last_peak}
\alias{mark_last_peak}
\title{Mark the location of the last maximum value in a vector}
\usage{
mark_last_peak(vec, threshold = NULL)
}
\arguments{
\item{vec}{(Numeric) A vector.}

\item{threshold}{(Numeric or NULL) The smallest acceptable peak value. NULL means no
threshold.}
}
\value{
A bool vector of the same length as \code{vec}, with the last maximum marked with
a TRUE.
}
\description{
Mark the location of the last maximum value in a vector
}
\section{Authors}{

\itemize{
\item Desi Quintans (\url{http://www.desiquintans.com})
}
}

\examples{
input <- c(1, 2, 3, 3, 1)
mark_last_peak(input, threshold = NULL)

#> [1] FALSE FALSE FALSE  TRUE FALSE

mark_last_peak(input, threshold = 4)

#> [1] FALSE FALSE FALSE FALSE FALSE

}
