% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_tools.R
\name{mark_longest_run}
\alias{mark_longest_run}
\title{Find and mark the longest run of TRUEs in a boolean vector}
\usage{
mark_longest_run(vec)
}
\arguments{
\item{vec}{(Logical) A vector.}
}
\value{
A new logical vector of the same length as \code{vec}, where the longest run of
TRUEs is marked with TRUE and all other values are marked FALSE. If there are two
runs of TRUE with equal length in \code{vec}, both will be reported in the results.
}
\description{
Find and mark the longest run of TRUEs in a boolean vector
}
\section{Authors}{

\itemize{
\item docendo discimus (\url{https://stackoverflow.com/users/3521006})
}
}

\section{Source}{

\url{https://stackoverflow.com/a/37447844}
}

\examples{
input <- c(T, T, F, F, F, F, F, T, T, T, T, T, T, F)
mark_longest_run(input)

#> [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE FALSE

}
