% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_tools.R
\name{match_in_substr}
\alias{match_in_substr}
\title{Match strings that have a needle near the start, end, or middle}
\usage{
match_in_substr(str, query, buffer = 0.25, from = "s", values = FALSE)
}
\arguments{
\item{str}{(Character) The vector to be searched.}

\item{query}{(Character) The regular expression to look for. If \code{length(query) > 1},
will be collapsed into a regular expression as \code{"(item1|item2|item3...)"}.}

\item{buffer}{(Numeric) The length of the substring to search. If given as a whole number,
it will be that many characters long. If given as a decimal number, it will be used
as a proportion of the length of each element in \code{str}, e.g. \code{buffer = 0.20} is 20\%
of each element's length.}

\item{from}{(Character) If \code{"start"} or \code{"s"} (default), the search will be done from
the start of each string. If \code{"end"} or \code{"e"}, it will be from the end. If \code{"middle"}
or \code{"m"}, the middle of the string will be searched.}

\item{values}{(Logical) If \code{FALSE} (default), returns a Logical vector of whether a match
was found in each element of \code{str}. If \code{TRUE}, returns a Character vector containing
only the elements of \code{str} that matched.}
}
\value{
A Logical vector if \code{values == FALSE} (default), or a Character vector
if \code{values == TRUE}.
}
\description{
Match strings that have a needle near the start, end, or middle
}
\examples{
sentences <- c(
    "The word 'needle' appears at the start of this sentence.",
    "But in this sentence, 'needle' doesn't.",
    "If 'needle' appears several times in a sentence, then we have a lot of needles!",
    "And in here, the word we want to find (needle) is near the middle of the sentence."
    )

# Within 20 characters of the Start of the string
match_in_substr(sentences, "needle", 20, "s")

#> [1] TRUE FALSE TRUE FALSE

# In the last 25\% of the string
match_in_substr(sentences, "needle", 0.25, "e")

#> [1] FALSE FALSE TRUE FALSE

# In the middle 1/3rd of the string
match_in_substr(sentences, "needle", 1/3, "m")

#> [1] FALSE FALSE FALSE TRUE

match_in_substr(sentences, "needle", 1/3, "m", values = TRUE)
#> [1] "And in here, the word we want to find (needle) is near the middle of the sentence."

}
