% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exports.R
\docType{data}
\name{material2014_colblind}
\alias{material2014_colblind}
\title{Google Material Design 2014 palette, with colour-blind conversions}
\format{
\subsection{\code{material2014_colblind}}{

A data frame with 280 rows and 6 columns:
\describe{
\item{name}{The colour's name.}
\item{normal}{Normal colour vision representation as hex colour.}
\item{deutan}{Deuteranopic colour conversion. Sees blues and yellows.}
\item{protan}{Protanopic conversion. Sees dimmer blues and yellows.}
\item{tritan}{Tritanopic conversion. Sees Red/pink and cyan.}
\item{achrom}{Achromatopic conversion. Sees greyscale.}
}
}
}
\source{
https://m2.material.io/design/color/the-color-system.html#tools-for-picking-colors
}
\usage{
material2014_colblind
}
\description{
These are the colours from Google's Material Design 2014 palette.
Colours are named according to the source website, e.g. \code{"deep_orange_50"}
through to \code{"deep_orange_a700"}. Colours with \code{"a"} before the number are
accent colours. #' More importantly, I provide simulated
colour-blind conversions of these colours that were computed with
\code{khroma:::anomalize()}. NOTE: I have padded this palette so that it displays
nicely with \code{show_colors(material2014_colblind$normal, n = 14)}. To unpad it,
remove rows where name is \code{NA}.
}
\keyword{datasets}
