% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_tools.R
\name{Mode}
\alias{Mode}
\title{Mode of a vector (numeric/character/factor)}
\usage{
Mode(x, break_ties = "no", na.rm = FALSE, ties = NULL, mean = NULL)
}
\arguments{
\item{x}{(Char/Numeric/Factor) A vector.}

\item{break_ties}{(Character) If more than one mode is found, how should the
tie be broken?
\itemize{
\item \code{"no"} or \code{FALSE}: Return a vector of all of the modes that were found.
\item \code{"random"}: Randomly choose one of the modes to return.
\item \code{"mean"}: Return the average of all of the modes (for numeric vectors).
\item \code{"first"}: Return the first mode found.
\item \code{"last"}: Return the last mode found.
\item \code{"median"}: Return the median value of all of the modes.
\item \code{"median l"} or \code{"median r"}: Return the mode to the left or right of the median.
\item \code{"NA"}: Return NA. Useful if you only want one clear winner.
}}

\item{na.rm}{(Logical) If \code{TRUE}, NAs will be silently removed.}

\item{ties}{Deprecated (2019-02-26). Use \code{break_ties} instead.}

\item{mean}{Deprecated (2019-02-26). Use \code{break_ties} instead.}
}
\value{
A vector of the mode value(s).
}
\description{
There is no built-in function to find the mode of something. This function
can find the mode of a numeric, character, or factor vector. By default it
will return multiple values in a multi-modal dataset, but there are several
methods of tie-breaking included.
}
\details{
If all values are unique, it will return \strong{all} of the values unless you
choose to break the tie.
}
\section{Authors}{

\itemize{
\item Ken Williams (\url{https://stackoverflow.com/users/169947/ken-williams})
\item jprockbelly (\url{https://stackoverflow.com/users/1502898/jprockbelly})
\item digEmAll (\url{https://stackoverflow.com/users/316644/digemall})
\item Desi Quintans (\url{http://www.desiquintans.com})
}
}

\section{Source}{

\url{https://stackoverflow.com/a/8189441/5578429}
}

\examples{
vec <- c(1, 2, 3, 4, 4, 4, 3, 3, NA, NA, NA)

Mode(vec, break_ties = "no")
#> [1]  3  4 NA

Mode(vec, break_ties = "no", na.rm = TRUE)
#> [1] 3 4

Mode(vec, break_ties = "mean", na.rm = FALSE)
#> [1] NA

Mode(vec, break_ties = "mean", na.rm = TRUE)
#> [1] 3.5

Mode(vec, break_ties = "median", na.rm = TRUE)
#> [1] 3

Mode(letters[1:4], break_ties = "no")
#> [1] "a" "b" "c" "d"

Mode(letters[1:4], break_ties = "median l")
#> "b"

Mode(letters[1:4], break_ties = "median r")
#> "c"

Mode(letters[1:4], break_ties = "random")
#> [1] "a"

}
