% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_tools.R
\name{na_in_row}
\alias{na_in_row}
\title{Count/proportion of \code{NA}s per dataframe row}
\usage{
na_in_row(df, ...)
}
\arguments{
\item{df}{(Dataframe) A dataframe.}

\item{...}{(Tidy-select) \code{dplyr}-style column selection.
See \code{\link[dplyr:dplyr_tidy_select]{dplyr::dplyr_tidy_select()}}. If empty, defaults to \code{dplyr::everything()}.}
}
\value{
The dataframe \code{df} with two new columns: \code{na_in_row_count} and \code{na_in_row_prop}.
}
\description{
Count/proportion of \code{NA}s per dataframe row
}
\section{Source}{

\itemize{
\item \url{https://stackoverflow.com/a/35444245/5578429}
}
}

\section{Authors}{

\itemize{
\item Desi Quintans (\url{http://www.desiquintans.com})
\item maloneypatr (\url{https://stackoverflow.com/users/2124146/maloneypatr})
}
}

\examples{
df <- data.frame(a = c(1, 2, NA, 4, NA), b = 1:5, c = c(NA, 2, 3, NA, NA))

df

#>     a b  c
#> 1   1 1 NA
#> 2   2 2  2
#> 3  NA 3  3
#> 4   4 4 NA
#> 5  NA 5 NA

# By default, looks for NAs in all columns
na_in_row(df)

#>    a b  c na_in_row_count na_in_row_prop
#> 1  1 1 NA               1      0.3333333
#> 2  2 2  2               0      0.0000000
#> 3 NA 3  3               1      0.3333333
#> 4  4 4 NA               1      0.3333333
#> 5 NA 5 NA               2      0.6666667

# Or use tidyselectors to choose columns. Here, looking for 
# NAs in all columns except `b`
na_in_row(df, -b)

#>    a b  c na_in_row_count na_in_row_prop
#> 1  1 1 NA               1            0.5
#> 2  2 2  2               0            0.0
#> 3 NA 3  3               1            0.5
#> 4  4 4 NA               1            0.5
#> 5 NA 5 NA               2            1.0

}
