% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_tools.R
\name{na_rm}
\alias{na_rm}
\title{Remove NAs from vectors, lists, matrices, and dataframes}
\usage{
na_rm(obj)
}
\arguments{
\item{obj}{(Vector/List/Matrix/Dataframe) The thing to remove \code{NA}s from.}
}
\value{
A copy of \code{obj} without NAs.
}
\description{
This is a wrapper around \code{stats::na_omit()} which hides its annoying printing
behaviour when applied to a vector.
}
\examples{
na_rm(c("a", NA, "b", NA, "c", "d"))

#> [1] "a" "b" "c" "d"


na_rm(data.frame(col1 = 1:5, col2 = c("a", NA, "c", "d", NA)))

#>   col1 col2
#> 1    1    a
#> 3    3    c
#> 4    4    d

}
