% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_tools.R
\name{normalize_colwise}
\alias{normalize_colwise}
\title{Normalise a matrix column-wise between 0 and 1.}
\usage{
normalize_colwise(mat)
}
\arguments{
\item{mat}{(Numeric) A numeric matrix.}
}
\value{
A numeric matrix. The largest value in the column will be \code{1}, and the other
values will be proportions compared to that.
}
\description{
Normalise a matrix column-wise between 0 and 1.
}
\section{Authors}{

\itemize{
\item endamaco (\url{https://stackoverflow.com/users/1558222/endamaco})
}
}

\section{Source}{

\url{https://stackoverflow.com/q/14282323/5578429}
}

\examples{
normalize_colwise(matrix(1:12, ncol = 3))

#>      [,1]  [,2]      [,3]
#> [1,] 0.25 0.625 0.7500000
#> [2,] 0.50 0.750 0.8333333
#> [3,] 0.75 0.875 0.9166667
#> [4,] 1.00 1.000 1.0000000

}
