% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_tools.R
\name{normalize_whole}
\alias{normalize_whole}
\title{Normalise a whole matrix between 0 and 1.}
\usage{
normalize_whole(mat, from_zero = FALSE)
}
\arguments{
\item{mat}{(Numeric) A matrix.}

\item{from_zero}{(Logical, default \code{FALSE}) If \code{FALSE}, the smallest value in the
matrix will be set to 0. If \code{TRUE}, the smallest value in the matrix will be
somewhere between 0 and 1.}
}
\value{
A matrix.
}
\description{
Normalise a whole matrix between 0 and 1.
}
\section{Authors}{

\itemize{
\item Desi Quintans (\url{http://www.desiquintans.com})
}
}

\examples{
normalize_whole(matrix(1:4, ncol = 2), from_zero = TRUE)

#>      [,1] [,2]
#> [1,] 0.25 0.75
#> [2,] 0.50 1.00

normalize_whole(matrix(1:4, ncol = 2), from_zero = FALSE)

#>           [,1]      [,2]
#> [1,] 0.0000000 0.6666667
#> [2,] 0.3333333 1.0000000

}
