% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_tools.R
\name{not.na}
\alias{not.na}
\title{Opposite of \code{is.na()}}
\usage{
not.na(vec)
}
\arguments{
\item{vec}{(Vector) Any vector.}
}
\value{
A logical vector of the same length as \code{vec}, with elements either
\code{TRUE} when the element isn't \code{NA}, or \code{FALSE} when it is \code{NA}.
}
\description{
Checks that vector elements are not \code{NA}. This is more readable and noticeable
than \code{!is.na(vec)}, and more compact than \code{is.na(vec) == FALSE}.
}
\section{Authors}{

\itemize{
\item Desi Quintans (\url{http://www.desiquintans.com})
}
}

\examples{
x <- c(1, NA, 2, 3, 4)

is.na(x)

## [1] FALSE  TRUE FALSE FALSE FALSE

not.na(x)

## [1]  TRUE FALSE  TRUE  TRUE  TRUE

}
