% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_tools.R
\name{not.nan}
\alias{not.nan}
\title{Opposite of \code{is.nan()}}
\usage{
not.nan(vec)
}
\arguments{
\item{vec}{(Vector) Any vector.}
}
\value{
A logical vector of the same length as \code{vec}, with elements either
\code{TRUE} when the element isn't \code{NaN}, or \code{FALSE} when it is \code{NaN}.
}
\description{
Checks that vector elements are not \code{NaN}. This is more readable and noticeable
than \code{!is.nan(vec)}, and more compact than \code{is.nan(vec) == FALSE}.
}
\section{Authors}{

\itemize{
\item Desi Quintans (\url{http://www.desiquintans.com})
}
}

\examples{
x <- c(1, NaN, 2, 3, 4)

is.nan(x)

## [1] FALSE  TRUE FALSE FALSE FALSE

not.nan(x)

## [1]  TRUE FALSE  TRUE  TRUE  TRUE

}
