% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_tools.R
\name{notin}
\alias{notin}
\alias{\%notin\%}
\title{Inverse match}
\usage{
x \%notin\% y
}
\arguments{
\item{x}{(Vector) The values to be matched. Long vectors (2^31 elements) are supported.}

\item{y}{(Vector) The values to be matched against. Long vectors are not supported.}
}
\value{
A logical vector of the same length as \code{x}, with \code{TRUE} if the element was
found in \code{y}, and \code{FALSE} if it was not.
}
\description{
Flags the elements of x that are not in y.
}
\section{Authors}{

\itemize{
\item R Core Team (\url{https://www.r-project.org/contributors.html})
\item Desi Quintans (\url{http://www.desiquintans.com})
}
}

\examples{
c(1, 4, 21, 7, -3) \%in\% 0:10
#> [1]  TRUE  TRUE FALSE  TRUE FALSE

c(1, 4, 21, 7, -3) \%notin\% 0:10
#> [1] FALSE FALSE  TRUE FALSE  TRUE

}
