% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_tools.R
\name{nth_word}
\alias{nth_word}
\title{Get the nth word from a vector of strings}
\usage{
nth_word(vec, n = 1, split = "(\\\\.|!|\\\\?){0,}( |$)")
}
\arguments{
\item{vec}{(Character) A vector of strings. It will be coerced to Character.}

\item{n}{(Integer) Which word will be retrieved? Positive integers count from
the start of the string, negative integers count from the end.}

\item{split}{(Character) A regular expression that matches word boundaries.
By default, it matches spaces and sentence-ending punctuation. See
\link[base]{strsplit}.}
}
\value{
A character vector that is the same length as \code{vec}. For \code{vec}
elements where no word was found (e.g. \code{NA} or an empty string \code{""}), \code{NA}
will be returned.
}
\description{
Get the nth word from a vector of strings
}
\section{Authors}{

\itemize{
\item Desi Quintans (\url{http://www.desiquintans.com})
}
}

\examples{
vec <- c("Lorem ipsum dolor",
         "sit amet, consectetur",
         "adipiscing elit, sed",
         "do eiusmod tempor",
         "incididunt ut labore",
         "et dolore magna",
         "aliqua.")

nth_word(vec, n = 1)

## [1] "Lorem"  "sit"  "adipiscing"  "do"  "incididunt"  "et"  "aliqua"

nth_word(vec, n = -1)

## [1] "dolor"  "consectetur"  "sed"  "tempor"  "labore"  "magna"  "aliqua" 

nth_word("Order_Family_Genus", n = 2, split = "_")  

## [1] "Family"

}
