% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_tools.R
\name{omit_dips}
\alias{omit_dips}
\title{Replace element with NA if it is less than the elements before and after it}
\usage{
omit_dips(vec)
}
\arguments{
\item{vec}{(Numeric) A vector.}
}
\value{
A Numeric vector.
}
\description{
Replace element with NA if it is less than the elements before and after it
}
\section{Authors}{

\itemize{
\item Desi Quintans (\url{http://www.desiquintans.com})
}
}

\examples{
tree_height_per_month <- c(1, 2, 3, 2, 4, 7, 7, 8)

omit_dips(tree_height_per_month)
#> [1]  1  2  3 NA  4  7  7  8

}
