% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_tools.R
\name{overwrite_df}
\alias{overwrite_df}
\title{Replace all matching values in a dataframe with something else}
\usage{
overwrite_df(df, find = "^(NA||\\\\s+|0|-+|_+)$", replace = "",
  replace_na = TRUE)
}
\arguments{
\item{df}{(Dataframe) A dataframe.}

\item{find}{(Character) A regex search pattern.}

\item{replace}{(Character) The string used to overwrite the matching cells.}

\item{replace_na}{(Logical) If \code{TRUE}, also overwrite R's built-in \code{NA} values.}
}
\value{
A dataframe.
}
\description{
Uses regex to match and replace cell values. This function is meant for formatting
printed data; it's often useful to blank out NAs or other values to minimise visual
clutter when you are reading a table. For actual data-tidying applications, it's
safer to use \code{dplyr::recode()} or \code{dplyr::recode_factor()}.
}
\details{
By default, this function will replace cells consisting of NAs, spaces, empty strings,
dashes, and underscores with an empty string.
}
\section{Authors}{

\itemize{
\item Desi Quintans (\url{http://www.desiquintans.com})
}
}

\examples{
test_df <-
    data.frame(stringsAsFactors = FALSE,
               name = c("insect1", "insect2", "insect3", "insect4", "insect5",
                        "insect6", "insect7", "insect8", "insect9", "insect10"),
               family = c("Belidae", "Belidae", " ", "Coccinelidae", NA, "Coccinelidae",
                          "Braconidae", "_", "-", "Curculionidae"),
               is_cool = c("TRUE", "TRUE", NA, "TRUE", "", "TRUE", "TRUE", "-", "_",
                           "TRUE")
    )

test_df
overwrite_df(test_df)

}
