% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{palette_distinct}
\alias{palette_distinct}
\title{1,022 visually-distinct colours}
\usage{
palette_distinct(n = 1022)
}
\arguments{
\item{n}{(Numeric) The number of colours to deliver. These are chosen sequentially from
the list.}
}
\value{
A character vector of hex colours.
}
\description{
There's so many colours, man. Many of these are not colourblind-safe! Many also have
poor contrast against each other, but they can still be differentiated as different
colours. Never rely on colour alone to differentiate data points; use shapes, opacity,
patterns, etc. Better yet, consider simplifying your graph so that it doesn't have so
many dimensions.
}
\details{
Code to plot all the colours can be found in the Examples section. You can even use a
colour picker to choose some and be assured that the colours can be differentiated.
}
\section{Authors}{

\itemize{
\item Tatarize (\url{https://stackoverflow.com/users/631911/tatarize})
\item Desi Quintans (<http://www.desiquintans.com)
}
}

\section{Source}{

\url{https://stackoverflow.com/a/12224359/5578429}
}

\examples{
# To see the colours (ordered left-to-right and top-to-bottom):
image(apply(matrix(1022:1, ncol = 73, nrow = 14, byrow = TRUE), 1, rev), col = palette_distinct())

# To get 4 colours:
palette_distinct(4)

#> [1] "#000000" "#FFFF00" "#1CE6FF" "#FF34FF"

}
