% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{palette_mrmrs}
\alias{palette_mrmrs}
\title{Adam Morse's palette of 16 web-safe colours}
\usage{
palette_mrmrs(n = NULL, random = FALSE, spaced = FALSE,
  alpha = NULL)
}
\arguments{
\item{n}{(Numeric) The number of colours to deliver.}

\item{random}{(Logical) If \code{FALSE} (default), you will get \code{n} colours in the order
they appear in the list. If \code{TRUE}, you will get \code{n} colours chosen randomly.}

\item{spaced}{(Logical) If \code{TRUE}, \code{n} colours will be chosen equally spaced within the
\code{col_list}, e.g. 3 colours from a list of 16 will return 1, 8, and 16th colours.
If \code{FALSE} (by default), colours will not be spaced out.}

\item{alpha}{(Numeric) A single decimal value that modifies the opacity of the colours.
\code{alpha = 0.65} makes all of the colours 65 percent opaque (35 percent transparent).
If \code{alpha = NULL}, no alpha channel is added and the colours are 100 percent opaque.}
}
\value{
A character vector of hex colours. If \code{alpha = NULL}, the colours will be in
RGB Hex format (e.g. #FFFF00). If \code{alpha} is not \code{NULL}, the colours will be in
RGBA Hex format (e.g. #FFFF00CB).
}
\description{
These are more attractive alternatives to the standard 16 default web-safe colours.
They have good contrast against each other, and the colour list is named.
}
\section{Authors}{

\itemize{
\item Adam Morse (\url{http://mrmrs.cc/})
\item Desi Quintans (<http://www.desiquintans.com)
}
}

\section{Source}{

\url{https://clrs.cc/}
}

\examples{
# To see all of the colours (ordered left-to-right and top-to-bottom):
show_colours(palette_mrmrs())

# The colours in the list are named
names(palette_mrmrs())

#> [1] "navy"    "blue"    "aqua"    "teal"    "olive"   "green"   "lime"    "yellow"
#> [9] "orange"  "red"     "maroon"  "fuchsia" "purple"  "black"   "gray"    "silver"

# To get the first 4 colours:
palette_mrmrs(4)

#>      navy      blue      aqua      teal
#> "#001F3F" "#0074D9" "#7FDBFF" "#39CCCC"

# To pick 4 colours randomly:
palette_mrmrs(4, random = TRUE)

#>    orange    maroon     olive   fuchsia
#> "#FF851B" "#85144B" "#3D9970" "#F012BE"

# To make the colours 75 percent opaque:
palette_mrmrs(4, random = TRUE, alpha = 0.75)

#>      orange      maroon       olive     fuchsia
#> "#FF851BBF" "#85144BBF" "#3D9970BF" "#F012BEBF"

}
