% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{palette_picked}
\alias{palette_picked}
\title{14 hand-picked distinct colours}
\usage{
palette_picked(n = NULL, random = FALSE, alpha = NULL)
}
\arguments{
\item{n}{(Numeric) The number of colours to deliver.}

\item{random}{(Logical) If \code{FALSE} (default), you will get \code{n} colours in the order
they appear in the list. If \code{TRUE}, you will get \code{n} colours chosen randomly.}

\item{alpha}{(Numeric) A single decimal value that modifies the opacity of the colours.
\code{alpha = 0.65} makes all of the colours 65 percent opaque (35 percent transparent).
If \code{alpha = NULL}, no alpha channel is added and the colours are 100 percent opaque.}
}
\value{
A character vector of hex colours. If \code{alpha = NULL}, the colours will be in
RGB Hex format (e.g. #FFFF00). If \code{alpha} is not \code{NULL}, the colours will be in
RGBA Hex format (e.g. #FFFF00CB).
}
\description{
The palette created by \code{desiderata::palette_distinct()} has a lot of colours that are
either so dark or so light that it's difficult to differentiate them next to each
other. In addition, many of the colours are affected by adjacency effects where
putting an intermediate colour between them makes them look the same.
}
\details{
I went through the preview plots manually, randomising the order of colours and
deleting colours that were visually similar until I ended up with a list of colours
that were easy to differentiate.
}
\section{Authors}{

\itemize{
\item Desi Quintans (<http://www.desiquintans.com)
}
}

\section{Source}{

\url{https://stackoverflow.com/a/12224359/5578429}
}

\examples{
# To see all of the colours (ordered left-to-right and top-to-bottom):
show_colours(palette_picked())

# The colours in the list are named
names(palette_picked())

#>  [1] "darkslateblue" "darkseagreen"  "darkorchid"    "yellow"        "magenta"       "plum"
#>  [7] "goldenrod"     "orangered"     "darkorange"    "skyblue"       "palegreen"     "dodgerblue"
#> [13] "chartreuse"    "grey10"

# To get the first 4 colours:
palette_picked(4)

#> darkslateblue  darkseagreen    darkorchid        yellow
#>     "#004754"     "#00AE7E"     "#7E2DD2"     "#FFE502"

# To pick 4 colours randomly:
palette_picked(4, random = TRUE)

#> dodgerblue   darkorchid darkseagreen    orangered
#>  "#3B5DFF"    "#7E2DD2"    "#00AE7E"    "#BE0028"

# To make the colours 75 percent opaque:
palette_picked(4, random = TRUE, alpha = 0.75)

#>  darkorange   goldenrod  darkorchid  chartreuse
#> "#FF6832BF" "#FEC96DBF" "#7E2DD2BF" "#4FC601BF"

}
