% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_tools.R
\name{pctin}
\alias{pctin}
\alias{\%pctin\%}
\title{Percentage of matching elements between two vectors}
\usage{
x \%pctin\% y
}
\arguments{
\item{x}{(Vector) The values to be matched.}

\item{y}{(Vector) The values to be matched against.}
}
\value{
The percentage of elements in x that are present in y.
}
\description{
Find the percentage of elements in x that are present in y.
}
\section{Authors}{

\itemize{
\item GSee (\url{https://stackoverflow.com/users/967840/gsee})
}
}

\section{Source}{

\url{https://stackoverflow.com/a/13830068/5578429}
}

\examples{
c(1, 4, 21, 7, -3) \%in\% 0:10
#> [1]  TRUE  TRUE FALSE  TRUE FALSE

c(1, 4, 21, 7, -3) \%pctin\% 0:10
#> [1] 0.6

}
