% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{plot_arrange}
\alias{plot_arrange}
\title{Arrange base R plots on a grid}
\usage{
plot_arrange(..., nrow, ncol, par_args = character(0))
}
\arguments{
\item{...}{(Expressions) Functions that create plots.}

\item{nrow}{(Integer) Number of rows to arrange plots into.}

\item{ncol}{(Integer) Number of columns to arrange plots into.}

\item{par_args}{(List) A \code{list()} of arguments to pass into \code{par()}.}
}
\value{
Plots arranged in a grid.
}
\description{
This is similar to \code{gridExtra::grid.arrange()}, but it works on base R
plots instead of ggplot objects.
}
\section{Authors}{

\itemize{
\item Desi Quintans (\url{http://www.desiquintans.com})
}
}

\examples{
plot_arrange(plot(Sepal.Length ~ Sepal.Width,  data = iris),
             plot(Petal.Length ~ Petal.Width,  data = iris),
             plot(Petal.Length ~ Sepal.Length, data = iris),
             plot(Petal.Width  ~ Sepal.Width,  data = iris),
             nrow = 2, ncol = 2)

}
