% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{quick_lm}
\alias{quick_lm}
\title{Fit and plot a two-term linear model quickly}
\usage{
quick_lm(formula, data, ...)
}
\arguments{
\item{formula}{(Formula) A two-term formula for both the linear model and the plot.}

\item{data}{(Dataframe) The dataframe to use for fitting and plotting.}

\item{...}{(Args) Arguments that will be passed to \code{plot()}.}
}
\value{
A plot that shows a fit line and lists intercept, slope, and adjusted r-squared.
}
\description{
Fit and plot a two-term linear model quickly
}
\section{Authors}{

\itemize{
\item Desi Quintans (\url{http://www.desiquintans.com})
}
}

\examples{
quick_lm(Petal.Length ~ Sepal.Length, iris)

}
