% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exports.R
\docType{data}
\name{random_integers}
\alias{random_integers}
\title{10,000 random integers between 0 and 100}
\description{
This list of numbers was generated by random.org's Integer Generator. It is
provided in the same order that the numbers were generated in.
}
\details{
This is meant for use as a column of "true-random" numbers for applications like
Random Forest variable importance, where one strategy for identifying good
features is to include a column of random values so that any feature less
important than these random values can be discarded.
}
\section{Author}{

\itemize{
\item Dr. Mads Haahr (\url{https://www.random.org/mads/})
}
}

\section{Source}{

Haahr, M. (2019, July 28). RANDOM.ORG: True Random Number Service. Retrieved from
https://www.random.org/integers/. Accessed 2019-07-28 23:18:13 UTC.
}

\examples{
# data(random_integers)

# head(random_integers)

#> [1] 93 18 19 91 53 54


#> data(random_integers)

}
