% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{rcols_as_hex}
\alias{rcols_as_hex}
\title{Convert R's built-in named colours to hex codes}
\usage{
rcols_as_hex(vec = NULL, distinct = FALSE)
}
\arguments{
\item{vec}{(Character) A character vector of R's built-in named colours. If left \code{NULL},
the entire contents of \code{grDevices::colours()} will be used.}

\item{distinct}{(Logical) Should only distinct colours be returned? Distinct colours do
not share the same (0:255)^3 RGB space, and do not have the same hex code. Defaults
to \code{FALSE} so that you don't get unexpectedly-short vectors.}
}
\value{
A named character vector of hex colours that were converted from R's built-in
names. The names are the original R colour names that were provided in \code{vec}.
}
\description{
Convert R's built-in named colours to hex codes
}
\examples{
rcols_as_hex(c("tomato", "steelblue"))
#>    tomato steelblue
#> "#FF6347" "#4682B4"

rcols_as_hex()
#>     white    aliceblue    antiquewhite    ... for all colours in colours()
#> "#FFFFFF"    "#F0F8FF"    "#FAEBD7"       ... for all colours in colours()

rcols_as_hex(distinct = TRUE)
#>     white    aliceblue    antiquewhite    ... for all in colours(distinct = TRUE)
#> "#FFFFFF"    "#F0F8FF"    "#FAEBD7"       ... for all in colours(distinct = TRUE)

rcols_as_hex(c("snow", "snow1"))
#>      snow     snow1
#> "#FFFAFA" "#FFFAFA"

rcols_as_hex(c("snow", "snow1"), distinct = TRUE)
#>      snow
#> "#FFFAFA"

}
