% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_tools.R
\name{report_diff}
\alias{report_diff}
\title{Report pairwise differences between two vectors}
\usage{
report_diff(l, r, full = TRUE)
}
\arguments{
\item{l}{(Vector) Any vector.}

\item{r}{(Vector) Any vector.}

\item{full}{(Logical) If \code{TRUE} (default), adds counts of each specific mismatch to the
dataframe output. If \code{FALSE}, only shows the first two rows of the output (the overall
count of matches and mismatches).}
}
\value{
A dataframe.
}
\description{
Compares each element of two vectors to each other and reports the number of total
matches and mismatches, as well as a count for each unique mismatch.
}
\section{Authors}{

\itemize{
\item Desi Quintans (\url{http://www.desiquintans.com})
}
}

\examples{
v1 <- c(1, 2, 3, 4, 5, NA, NA)
v2 <- c(1, 2, 5, 4, 3, NA, 7)

report_diff(v1, v2)

#>         pairs count
#> 1      l == r     4
#> 2      l != r     3
#> 3  `3` != `5`     1
#> 4  `5` != `3`     1
#> 5 `NA` != `7`     1

report_diff(v1, v2, full = FALSE)

#>    pairs count
#> 1 l == r     4
#> 2 l != r     3

}
