% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{rotate_axis_text}
\alias{rotate_axis_text}
\alias{rotate_x_text}
\alias{rotate_y_text}
\title{Rotate and align ggplot2 axis tick labels}
\usage{
rotate_x_text(angle = -90, align = 0, valign = 0.25)

rotate_y_text(angle = 90, align = 0.5, valign = 0.5)
}
\arguments{
\item{angle}{(Numeric) Rotation angle. \code{-90} rotates 90 degrees clockwise, \code{90} rotates
90 degrees anti-clockwise.}

\item{align}{(Numeric) Horizontal alignment. \code{0} left-aligns, \code{1}
right-aligns, and \code{0.5} centers.}

\item{valign}{(Numeric) Vertical alignment. \code{0} top-aligns, \code{1} bottom-aligns,
and \code{0.5} centers.}
}
\value{
A ggplot2 theme object.
}
\description{
Rotate and align ggplot2 axis tick labels
}
\examples{
# library(ggplot2)
# ggplot(mpg, aes(manufacturer, cty)) + geom_boxplot() +
#     rotate_x_text() +
#     rotate_y_text()

#> A plot with X axis labels rotated 90 degrees clockwise and left-aligned to the
#> bottom edge of the plot area, and the Y axis labels rotated 90 degrees counter-
#> clockwise. All axis labels are centered on the tick marks.

}
