% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{rotate_facet_text}
\alias{rotate_facet_text}
\alias{rotate_x_facet_text}
\alias{rotate_y_facet_text}
\title{Rotate and align ggplot2 facet labels}
\usage{
rotate_x_facet_text(angle = 45, align = 0, valign = 0.25)

rotate_y_facet_text(angle = 45, align = 0, valign = 0.25)
}
\arguments{
\item{angle}{(Numeric) Rotation angle. \code{-90} rotates 90 degrees clockwise, \code{90} rotates
90 degrees anti-clockwise.}

\item{align}{(Numeric) Horizontal alignment. \code{0} left-aligns, \code{1}
right-aligns, and \code{0.5} centers.}

\item{valign}{(Numeric) Vertical alignment. \code{0} top-aligns, \code{1} bottom-aligns,
and \code{0.5} centers.}
}
\value{
A ggplot2 theme object.
}
\description{
Rotate and align ggplot2 facet labels
}
