% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_tools.R
\name{round_to}
\alias{round_to}
\title{Round numbers to a fixed number of decimal places}
\usage{
round_to(x, digits = 2, ...)
}
\arguments{
\item{x}{(Numeric) A vector of numbers.}

\item{digits}{(Numeric) Number of decimal places to keep.}
}
\value{
A numeric vector.
}
\description{
Round numbers to a fixed number of decimal places
}
\section{Authors}{

\itemize{
\item Jeromy Anglim (\url{https://stackoverflow.com/users/180892/jeromy-anglim})
}
}

\section{Source}{

\url{https://stackoverflow.com/a/12135122}
}

\examples{
vec <- c(1.739006, 2, -1.4, 1.05, 1.90, 3.826)
round_to(vec, digits = 3)
#> [1] "1.739"  "2.000"  "-1.400" "1.050"  "1.900"  "3.826" 

}
