% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime_tools.R
\name{round_to_duration}
\alias{round_to_duration}
\title{Round a duration of days to other units}
\usage{
round_to_duration(num, from, to, nearest = NULL, dir = NULL)
}
\arguments{
\item{num}{(Numeric/Integer) Time durations you want to convert.}

\item{from}{(Character) The units of time that \code{num} is in, e.g. \code{"days"}.}

\item{to}{(Character) The units of time to convert to, e.g. \code{"weeks"}.}

\item{nearest}{(Numeric/Integer or \code{NULL}) The closest number to round the
results to, e.g. \code{1} to round to integers. If \code{NULL}, no rounding is done.}

\item{dir}{(Character or \code{NULL}) Controls the rounding function used if
\code{nearest} is not \code{NULL}. If omitted or \code{"both"}, the \code{round()} function is
used. If \code{"up"}, \code{ceiling()} is used. If \code{"down"}, \code{floor()} is used.}
}
\value{
A numeric vector.
}
\description{
Convert a duration of days into approximate months or years, rounding it to a
nearest value if desired. Note that time durations larger than the week
are only approximate.
}
\section{Authors}{

\itemize{
\item Desi Quintans (\url{http://www.desiquintans.com})
}
}

\examples{
round_to_duration(134, "days", "weeks")
#> [1] 19.14286

# Note that this is the same as a naive calculation; if you were converting 
# days to months, then it's just days divided by 30.4375 (the average number 
# of days per month).
round_to_duration(134, "days", "months")
#> [1] 4.402464
134 / 30.4375
#> [1] 4.402464

# You can round the output too. This is useful if you want to express a time
# duration like, "1.5 months".
round_to_duration(134, "days", "weeks", nearest = 0.5, dir = "up")
#> [1] 19.5
round_to_duration(134, "days", "weeks", nearest = 0.5, dir = "down")
#> [1] 19

}
