% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_tools.R
\name{round_to_nearest}
\alias{round_to_nearest}
\title{Round numbers to a nearest "pretty" value}
\usage{
round_to_nearest(num, to, dir = NULL)
}
\arguments{
\item{num}{(Numeric) A vector of numbers.}

\item{to}{(Numeric) What interval should \code{num} be rounded to?}

\item{dir}{(Optional Char: \code{"up"}, \code{"down"}, \code{"both"}, or omitted) Controls the rounding
function used. If omitted or \code{"both"}, the \code{round()} function is used. If \code{"up"}, \code{ceiling()} is
used. If \code{"down"}, \code{floor()} is used.}
}
\value{
A numeric vector.
}
\description{
It's sometimes useful to round numbers to a predictable range, e.g. when preparing
graph axes or perhaps for filtering or presenting data. This function can round numbers
to the nearest, say, 0.5.
}
\section{Authors}{

\itemize{
\item Koshke (\url{https://stackoverflow.com/users/314020/kohske})
}
}

\section{Source}{

\url{https://stackoverflow.com/a/8665247/5578429}
}

\examples{
vec <- c(1.739006, 2, -1.4, 1.05, 1.90, 3.826)
round_to_nearest(vec, 0.5)
#> [1]  1.5  2.0 -1.5  1.0  2.0  4.0

}
