% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_tools.R
\name{round_to_places}
\alias{round_to_places}
\title{Round a number to a fixed decimal place length}
\usage{
round_to_places(num, places = 2)
}
\arguments{
\item{num}{(Numeric) A vector of numbers.}

\item{places}{(Numeric) The number of decimal places to keep.}
}
\value{
A Character vector of rounded numbers.
}
\description{
R's \code{round()} and \code{signif()} functions drop trailing zeros because they are not
significant, but it's sometimes important to keep them for formatting i.e. tables and
graphs. This function does that.
}
\section{Authors}{

\itemize{
\item Kristine Yu (\url{https://kmyu.wordpress.com})
}
}

\section{Source}{

\url{https://kmyu.wordpress.com/2011/01/11/formatting-numbers-for-printing-in-r-rounding-and-trailing-zeroes/}
}

\examples{
vec <- c(1.739006, 2, -1.4, 1.05, 1.90, 3.826)
rounded_vec <- round_to_places(vec, 2)

str(rounded_vec)
#> chr [1:6] "1.70" "2.00" "-1.40" "1.00" "1.90" "3.80"

}
