% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_tools.R
\name{same_cols}
\alias{same_cols}
\title{Given two dataframes, which columns appear in both of them?}
\usage{
same_cols(l, r)
}
\arguments{
\item{l}{(Dataframe) A dataframe whose column names to compare.}

\item{r}{(Dataframe) A dataframe whose column names to compare.}
}
\value{
A Character vector with the names of the columns that appear
in both \code{l} and \code{r}. The order of \code{l} and \code{r} doesn't matter for \code{same_cols()}, but
it does for \code{diff_cols()}.
}
\description{
Given two dataframes, which columns appear in both of them?
}
\section{Authors}{

\itemize{
\item Desi Quintans (\url{http://www.desiquintans.com})
}
}

\examples{
iris1 <- iris[, 1:3]
colnames(iris1)

#> [1] "Sepal.Length" "Sepal.Width"  "Petal.Length"

iris2 <- iris[, 2:5]
colnames(iris2)
#> [1]                "Sepal.Width"  "Petal.Length" "Petal.Width"  "Species"


same_cols(iris1, iris2)
#> [1] "Sepal.Width"  "Petal.Length"

}
