% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_tools.R
\name{save_a4}
\alias{save_a4}
\title{Save a ggplot as an A4 image}
\usage{
save_a4(path, basename, plot = ggplot2::last_plot(), portrait = FALSE)
}
\arguments{
\item{path}{(Character) The output folder. It will be created if it does not
exist. One sub-directory (\verb{/png/}) will also be created if it doesn't exist.}

\item{basename}{(Character) Basename of the image, without path or extension.
Always saved as a \code{.png}.}

\item{plot}{(ggplot) The plot object to save. By default, saves the most
recent ggplot.}

\item{portrait}{(Logical) \code{FALSE} by default, which outputs in landscape.
If \code{TRUE}, flips the page dimensions to output in portrait.}
}
\value{
Returns \code{TRUE} invisibly if successful.
}
\description{
The dimensions are for the pixel size of an A4 page at 300 DPI.
This is appropriate for inserting into Word without much detail loss, and
with appropriate text and element sizes.
}
\examples{
# qplot(mpg, wt, data = mtcars)
# save_a4("_test", "x_mpg y_wt")

#> Wrote 'ggplot2::last_plot' to '_test/png/x_mpg y_wt.png'.

}
