% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific_calculations.R
\name{se_mean}
\alias{se_mean}
\title{Standard error of the mean}
\usage{
se_mean(vec, na.rm = FALSE)
}
\arguments{
\item{vec}{(Numeric) A vector.}

\item{na.rm}{(Logical) If `TRUE`, remove `NA`s from `vec`.}
}
\value{
The standard error of the mean of `vec`.
}
\description{
The standard error of the mean is how the estimated mean changes with multiple
measurements (i.e. how far away the mean of each sampling event or observation is from
the true population mean). SE drops as sample size grows because as you take more
mreasurements, the sampling means cluster more closely to the true mean.
}
\examples{
se_mean(c(1, 2, 3, 4, NA_integer_))
#> NA

se_mean(c(1, 2, 3, 4, NA_integer_), na.rm = TRUE)
#> [1] 0.6454972
}
