% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_tools.R
\name{set_seed_any}
\alias{set_seed_any}
\title{Seed the random number generator with a character string (or any object)}
\usage{
set_seed_any(seed)
}
\arguments{
\item{seed}{(Any) Any object.}
}
\value{
\code{NULL}.
}
\description{
Set the random number generator's seed using a string if you want to be extra cute and
use your cats' names like I do. This function can actually generate a seed from any R
object, so you could even feed it a whole dataframe if you felt like it. (Requires the
\href{https://cran.r-project.org/web/packages/digest/index.html}{'digest'} package.)
}
\section{Authors}{

\itemize{
\item Ben Bolker (\url{https://stackoverflow.com/users/190277/ben-bolker})
}
}

\section{Source}{

\url{https://stackoverflow.com/a/10913336/5578429}
}

\examples{
# set_seed_any("Snake... Do you think love can bloom, even on a battlefield?")

# set_seed_any(iris)

}
