% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_tools.R
\name{Show}
\alias{Show}
\title{Pipeline- and knit-friendly \code{View()}}
\usage{
Show(x)
}
\arguments{
\item{x}{(Object) An object to \code{View()}.}
}
\value{
Runs \code{View()} on the object if R is running interactively. Finally,
invisibly returns \code{x}.
}
\description{
This function can be safely inserted into a pipeline while it is being developed.
It will run \code{View()} on the object only if R is running interactively. It returns
its input so that the pipeline can continue running.
}
\examples{
\dontrun{
Show(iris)
}

}
