% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{show_colours}
\alias{show_colours}
\title{Preview a list of colours as a grid}
\usage{
show_colours(col_list, asp = NA, main = NULL)
}
\arguments{
\item{col_list}{(Character) A vector of colours in RGB Hex or RGBA Hex format.}

\item{asp}{(Numeric or \code{NA}) The aspect ratio of the image. \code{asp = 1} produces a square
plot with square cells. If \code{NA}, \code{plot.window}'s default will be used.}

\item{main}{(Character or \code{NULL}) Leave this as \code{NULL} to generate a title. Supply a
string to define your own title.}
}
\value{
A graphic
}
\description{
Preview a list of colours as a grid
}
\examples{
show_colours(colours(distinct = TRUE))

}
