% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_tools.R
\name{shush}
\alias{shush}
\title{Suppress all console printing (cat, print, warning, message)}
\usage{
shush(x)
}
\arguments{
\item{x}{(Expression) An expression, usually a call to a function.}
}
\value{
The returned value of the expression.
}
\description{
Sometimes developers leave debugging messages in their packages or, infuriatingly,
choose to output messages by using print() or cat() instead of message() or warning()
like they're supposed to. This function suppresses them to remove that clutter.
}
\section{Authors}{

\itemize{
\item Desi Quintans (\url{http://www.desiquintans.com})
\item hplieninger (\url{https://stackoverflow.com/users/2563804/hplieninger})
}
}

\section{Source}{

\url{https://stackoverflow.com/a/48503375/5578429}
}

\examples{
loud_mean <- function(x) {
    print("This is from print().")
    cat("This is from cat().\n")
    message("This is from message().")
    warning("This is from warning().")
    mean(x)
}

loud_mean(1:100)

#> [1] "This is from print()."
#> This is from cat().
#> This is from message().
#> [1] 50.5
#> Warning message:
#>     In loud_mean(1:100) : This is from warning().

shush(loud_mean(1:100))

#> [1] 50.5

# magrittr pipelines will also work.
# shush(loud_mean(1:100) \%>\% sqrt())
#> [1] 7.106335

}
