% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_tools.R
\name{sift}
\alias{sift}
\title{Fuzzily search dataframe columns and labels with regex}
\usage{
sift(df, query, dist = 0.1)
}
\arguments{
\item{df}{(Dataframe) The dataframe}

\item{query}{(Character) A string (or regular expression) to search for.}

\item{dist}{(Numeric) Maximum string distance. Either an integer or a double; see
\code{max.distance} in \code{base::agrep()}. Set to \code{0} to force an exact match.}
}
\value{
Prints out matching name & label combinations, and invisibly returns the same.
}
\description{
When working with dataframes hundreds of columns wide, it can be hard to find which
variable contains a particular measurement. This function lets you fuzzily search a
dataframe's column names and labels (the \code{attr()} named \code{"label"}) for a match.
}
\section{Authors}{

\itemize{
\item Desi Quintans (\url{http://www.desiquintans.com})
}
}

\examples{

sift(dplyr::starwars, "col")

#> eye_color
#> hair_color
#> skin_color
#> vehicles

sift(dplyr::starwars, "col", dist = 0)

#> eye_color
#> hair_color
#> skin_color

}
