% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_tools.R
\name{sort_cols}
\alias{sort_cols}
\title{Sort columns in a dataframe}
\usage{
sort_cols(df, ..., decreasing = FALSE)
}
\arguments{
\item{df}{(Dataframe) A dataframe.}

\item{...}{(Column names) If you want to manually position columns \emph{after} they are
sorted, provide unquoted column names here. The columns in \code{...} will be placed
first in the dataframe, and then all other unlisted columns will be placed after.}

\item{decreasing}{(Logical) If \code{FALSE}, sort columns from A-Z and 0-9. If \code{TRUE}, sort
in reverse.}
}
\value{
A copy of \code{df} with reordered columns.
}
\description{
Sorts the columns of a dataframe, and then allows you to pull columns to the start of
the dataframe by name.
}
\section{Authors}{

\itemize{
\item Desi Quintans (\url{http://www.desiquintans.com})
}
}

\examples{
colnames(iris)

#> [1] "Sepal.Length" "Sepal.Width"  "Petal.Length" "Petal.Width"  "Species"

sorted <- sort_cols(iris)
colnames(sorted)

#> [1] "Petal.Length" "Petal.Width"  "Sepal.Length" "Sepal.Width"  "Species"

reverse <- sort_cols(iris, decreasing = TRUE)
colnames(reverse)

#> [1] "Species"  "Sepal.Width"  "Sepal.Length" "Petal.Width"  "Petal.Length"

manual <- sort_cols(iris, Species)
colnames(manual)

#> [1] "Species" " Petal.Length" "Petal.Width"  "Sepal.Length" "Sepal.Width"

}
