% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_funs.R
\name{split_n}
\alias{split_n}
\title{Split a vector into n groups}
\usage{
split_n(vec, g, balance = TRUE)
}
\arguments{
\item{vec}{(Numeric or Character) A vector. It will be sorted with \code{sort()} for grouping,
but returned in the original order.}

\item{g}{(Integer) The maximum number of groups to split \code{vec} into (for some group
sizes, the maximum will not be used).}

\item{balance}{(Logical) If \code{TRUE}, try to have equal numbers of observations per group.}
}
\value{
A vector of integers.
}
\description{
This function was deprecated on 2022-03-22. Use \code{assign_groups()} instead.
}
\examples{
testvec <- c(0.7685, 0.4116, 0.1416, 0.8450, 0.9021, 0.4965, 0.8341, 0.0438)

order(testvec)
#> [1] 8 3 2 6 1 7 4 5

split_n(testvec, 4)
#> [1] 4 2 1 3 1 4 2 3

split_n(testvec, 7, balance = TRUE)  # The range of groups is limited
#> [1] 4 2 1 3 1 4 2 3

split_n(testvec, 7, balance = FALSE)  # Try to use the whole range
#> [1] 7 2 1 5 1 6 3 4

split_n(testvec, 3)  # Sometimes unbalanced groups are inevitable
#> [1] 3 1 1 2 1 3 2 2

}
