% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_tools.R
\name{split_size}
\alias{split_size}
\title{Split a vector into chunks of size \code{n}}
\usage{
split_size(vec, size)
}
\arguments{
\item{vec}{(Vector) Any vector.}

\item{size}{(Integer) The number of entries per chunk.}
}
\value{
A list of dynamic length, where each entry is a chunk from \code{vec} of
length 1 up to \code{size}.
}
\description{
In contrast to \code{desiderata::assign_groups()} which splits a vector into \code{n}
chunks (possibly with a different number of entries per chunk),
\code{desiderata::split_size()} splits a vector into an arbitrary number of chunks
as long as each chunk has \code{n} or fewer entries inside it.
}
\section{Authors}{

\itemize{
\item Harlan (\url{https://stackoverflow.com/users/135944/harlan})
\item dfrankow (\url{https://stackoverflow.com/users/34935/dfrankow})
}
}

\section{Source}{

\url{https://stackoverflow.com/a/3321659/5578429}
}

\examples{
split_size(letters, 5)

## $`1`
## [1] "a" "b" "c" "d" "e"
## 
## $`2`
## [1] "f" "g" "h" "i" "j"
## 
## $`3`
## [1] "k" "l" "m" "n" "o"
## 
## $`4`
## [1] "p" "q" "r" "s" "t"
## 
## $`5`
## [1] "u" "v" "w" "x" "y"
## 
## $`6`
## [1] "z"

}
