% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_tools.R
\name{str_coll}
\alias{str_coll}
\title{Collapse a vector into a string}
\usage{
str_coll(..., wrap = "", between = ",", unique = FALSE)
}
\arguments{
\item{...}{(...) Vectors that will be concatenated and coerced to Character.}

\item{wrap}{(Character) Placed at the left and right sides of each vector element.}

\item{between}{(Character) Placed between each element of the original vector(s).}

\item{unique}{(Logical) If \code{TRUE}, duplicate entries in \code{...} will be removed.}
}
\value{
A string.
}
\description{
Take the elements of multiple vectors and concatenate them into one long string.
}
\section{Authors}{

\itemize{
\item Desi Quintans (\url{http://www.desiquintans.com})
}
}

\examples{
\dontrun{
vec <- c("Gum", "Rock", "Pen", "Bottle", "Pen")

str_coll(vec)
#> [1] "Gum,Rock,Pen,Bottle,Pen"

str_coll(vec, wrap = "-", between = ", ")
#> [1] "-Gum-, -Rock-, -Pen-, -Bottle-, -Pen-"

str_coll(vec, wrap = "-", between = ", ", unique = TRUE)
#> [1] "-Gum-, -Rock-, -Pen-, -Bottle-"
}

}
