% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_tools.R
\name{str_rev}
\alias{str_rev}
\title{Reverse every element of a vector}
\usage{
str_rev(..., USE.NAMES = FALSE)
}
\arguments{
\item{...}{(Vectors) Vectors that will be coerced to Character and joined together.}

\item{USE.NAMES}{(Logical) Should the output be a named vector, or unnamed?}
}
\value{
A Character vector where every element from the original vectors has been
reversed, but the order of the elements within that vector is still the same.
NAs remain as NA.
}
\description{
This reverses the individual elements (makes words go backwards, for example),
but does not change the order of those elements in the vector.
}
\section{Authors}{

\itemize{
\item Kevin Ushey (\url{https://stackoverflow.com/users/1342082/kevin-ushey})
}
}

\section{Source}{

\url{https://stackoverflow.com/a/14029000/5578429}
}

\examples{
vec <- c("Beret", "Clipper", "Cornet", NA)
str_rev(vec)

#> [1] "tereB"   "reppilC" "tenroC"  NA

}
