% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_tools.R
\name{top_tail}
\alias{top_tail}
\title{First and last rows of a dataframe}
\usage{
top_tail(df, top = 3, tail = 3)
}
\arguments{
\item{df}{(Dataframe) A dataframe.}

\item{top}{(Integer) The number of rows to get from the start of \code{df}.}

\item{tail}{(Integer) The number of rows to get from the end of \code{df}.}
}
\value{
A dataframe.
}
\description{
First and last rows of a dataframe
}
\details{
\code{0} can be provided for the top and tail, in which case it will behave like
\code{head()} and \code{tail()} respectively.
}
\section{Authors}{

\itemize{
\item Desi Quintans (\url{http://www.desiquintans.com})
}
}

\examples{
# Returns 6 rows by default, just like head() does.
top_tail(iris)

#>     Sepal.Length Sepal.Width Petal.Length Petal.Width   Species
#> 1            5.1         3.5          1.4         0.2    setosa
#> 2            4.9         3.0          1.4         0.2    setosa
#> 3            4.7         3.2          1.3         0.2    setosa
#> 148          6.5         3.0          5.2         2.0 virginica
#> 149          6.2         3.4          5.4         2.3 virginica
#> 150          5.9         3.0          5.1         1.8 virginica

top_tail(iris, top = 1, tail = 2)

#>     Sepal.Length Sepal.Width Petal.Length Petal.Width   Species
#> 1            5.1         3.5          1.4         0.2    setosa
#> 149          6.2         3.4          5.4         2.3 virginica
#> 150          5.9         3.0          5.1         1.8 virginica

}
