% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific_calculations.R
\name{triangle_num}
\alias{triangle_num}
\title{nth Triangle Number (like factorial but with addition)}
\usage{
triangle_num(nums, coerce = round, ...)
}
\arguments{
\item{nums}{(Numeric) A vector of numbers.}

\item{coerce}{(Name/Symbol) The name of a function that will be used to coerce
\code{nums} into integers, e.g. \code{round} or \code{ceiling} or \code{floor}.}

\item{...}{(dots) Extra arguments sent to the function named in \code{coerce}.}
}
\value{
A Numeric vector where each element is the nth triangle number of the
original element.
}
\description{
nth Triangle Number (like factorial but with addition)
}
\section{Authors}{

\itemize{
\item Desi Quintans (\url{http://www.desiquintans.com})
}
}

\section{Source}{

\url{https://math.stackexchange.com/a/60581}
}

\examples{
triangle_num(1:10)

#>  [1]  1  3  6 10 15 21 28 36 45 55

}
