% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_tools.R
\name{try_seed}
\alias{try_seed}
\title{Generate random seeds to preview their effects}
\usage{
try_seed(expr)
}
\arguments{
\item{expr}{(Expression) An expression.}
}
\value{
The evaluated \code{expr}.
}
\description{
This should not be used in scripts, it is for interactive use only. It will throw an
error if it is run non-interactively (e.g. if it is found inside an Rmarkdown document
that is being knitted, or an R script that is being executed from the command line).
This is because \code{try_seed()} changes the random seed, which could affect the rest
of your script.
}
\details{
This function picks a random seed, announces what that seed is, and then uses it to
evaluate an expression. For example, if you are creating a network graph whose layout
is calculated from randomly-chosen starting positions, \code{try_seed()} lets you run that
plotting function over and over with a new seed each time, until you find a layout that
you would like to keep. At that point, you would copy the announced seed from the
console and manually \code{set.seed()} in your script.
}
\section{Authors}{
 - Desi Quintans (\url{http://www.desiquintans.com})
}

\examples{
try_seed(runif(5))

#> Seed is: 1915981367
#> [1] 0.29910233 0.79275922 0.04287227 0.51237626 0.10189918
#>

set.seed(1915981367)  # The announced seed
runif(5)

#> [1] 0.29910233 0.79275922 0.04287227 0.51237626 0.10189918

}
