% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_tools.R
\name{unique_n}
\alias{unique_n}
\title{Keep the first n unique elements in a vector}
\usage{
unique_n(vec, n, sort = "no")
}
\arguments{
\item{vec}{(Vector) Any vector.}

\item{n}{(Integer) The maximum number of elements to keep from \code{vec}. If \code{n} is
greater than the number of elements in \code{vec}, the entirety of \code{vec} will be
returned.}

\item{sort}{(Character) \code{"ascending"} (or \code{"asc"} or \code{"a"}) sorts \code{vec} in
ascending order (the default order of \code{sort()}). \code{"descending"} (or \code{"desc"}
or \code{"d"}) sorts in reverse. Use \code{"no"} (or \code{"n"}) to leave the elements of
\code{vec} in their original order.}
}
\value{
A vector of the same type as \code{vec}.
}
\description{
Keep the first n unique elements in a vector
}
\examples{
set.seed(12345)
lets <- sample(letters[1:10], 50, replace = TRUE)

unique_n(lets, 10, sort = "no")

## [1] "c" "j" "h" "b" "f" "g" "a" "d" "i" "e"

unique_n(lets, 8, sort = "asc")

## [1] "a" "b" "c" "d" "e" "f" "g" "h"

unique_n(lets, 5, sort = "desc")

## [1] "j" "i" "h" "g" "f"

unique_n(1:5, 20)  # Asking for 20 elements, but there are only 5 uniques.

## [1] 1 2 3 4 5

}
