% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_tools.R
\name{vec_to_regex}
\alias{vec_to_regex}
\title{Collapse vectors into a regex pattern}
\usage{
vec_to_regex(..., sep = "|", wrap = c("(", ")"))
}
\arguments{
\item{...}{(Vector) Vectors that will be coerced into character. Duplicated elements
from the vectors are removed.}

\item{sep}{(Char) Separator to use between elements of ...}

\item{wrap}{(Char) 2-element vector of text to wrap the pattern with. \code{wrap[1]} is
placed on the left side of the pattern, \code{wrap[2]} is placed on the right side.}
}
\value{
A character string where every element of \code{...} is collapsed with \code{sep}, and
wrapped with \code{wrap}.
}
\description{
Collapse vectors into a regex pattern
}
\section{Authors}{

\itemize{
\item Desi Quintans (\url{http://www.desiquintans.com})
}
}

\examples{
vec_to_regex(month.abb)
#> [1] "(Jan|Feb|Mar|Apr|May|Jun|Jul|Aug|Sep|Oct|Nov|Dec)"

vec_to_regex(letters[1:6], sep = "", wrap = c("([", "]+)"))
#> [1] "([abcdef]+)"

}
