% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_tools.R
\name{write_csv_rds}
\alias{write_csv_rds}
\title{Write a dataframe to a CSV and RDS}
\usage{
write_csv_rds(df, path, basename)
}
\arguments{
\item{df}{(Dataframe) A dataframe.}

\item{path}{(Character) The output folder. It will be created if it does not
exist. Two sub-directories (\verb{/csv/} and \verb{/rds/}) will also be created if
they don't exist.}

\item{basename}{(Character) The name of the resulting files, without a
path or extension.}
}
\value{
Returns \code{TRUE} invisibly on success.
}
\description{
Write a dataframe to a CSV and RDS
}
\examples{
# write_csv_rds(iris, "_output", "my_iris")

#> Wrote 'iris' to '_output/csv/my_iris.csv'.
#> Wrote 'iris' to '_output/rds/my_iris.rds'.

}
