% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exports.R
\docType{data}
\name{basic_colour_names}
\alias{basic_colour_names}
\title{The names of 197 common colours}
\description{
This is a character vector of common colour names, lowercased and sorted by word length
so that common colours like red and blue come earlier in the list.
}
\section{Author}{

\itemize{
\item Desi Quintans (@eco_desi)
}
}

\section{Source}{

\itemize{
\item \url{https://simple.wikipedia.org/wiki/Web_color}
\item \url{https://simple.wikipedia.org/wiki/List_of_colors}
}
}

\examples{

}
