% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_tools.R
\name{drop_empty_cols}
\alias{drop_empty_cols}
\title{Drop 'empty' columns in a dataframe}
\usage{
drop_empty_cols(df)
}
\arguments{
\item{df}{(Dataframe) A dataframe to filter.}
}
\value{
A copy of \code{df} with all empty columns removed.
}
\description{
A column is empty when every single row is NA, NULL, "", or 0.
}
\examples{
data <- data.frame(a = c(1, 2, 3),
                   b = c(0, 0, 0),
                   c = c(1, 1, 0),
                   d = c("", "", ""),
                   e = c("moo", "baa", "woof"))

#> a b c d    e
#> 1 0 1    moo
#> 2 0 1    baa
#> 3 0 0    woof

drop_empty_cols(data)

#> a c    e
#> 1 1  moo
#> 2 1  baa
#> 3 0 woof

}
