% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_funs.R
\name{find_dims}
\alias{find_dims}
\title{Find the nearest rectangle that will fit the number of entries in a vector}
\usage{
find_dims(vec, x = NULL)
}
\arguments{
\item{vec}{(Vector) A vector.}

\item{x}{(Integer) Optionally, the desired width in cells.}
}
\value{
A named numeric vector with 2 entries: \code{x} and \code{y} describing the number of
columns and rows respectively.
}
\description{
This function is used by \code{show_colours()} to find the dimensions of a grid (as close
to square as possible) that will fit all of the entries in a vector of colours.
}
\examples{
\dontrun{
find_dims(1:12)

#> x y
#> 3 4

find_dims(1:12, 2)

#> x y
#> 2 6
}

}
