% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_funs.R
\name{howmany}
\alias{howmany}
\alias{howmany.default}
\alias{howmany.data.frame}
\alias{howmany.table}
\title{Count the number of unique values}
\usage{
howmany(x, ...)

\method{howmany}{default}(x, ...)

\method{howmany}{data.frame}(x, ...)

\method{howmany}{table}(x, ...)
}
\arguments{
\item{x}{(Object) The object to count.}

\item{...}{\verb{table:} Column names or positions that are passed to \code{dplyr::select()}.}
}
\description{
Count the number of unique values
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Returns the number of unique values in \code{x}.

\item \code{data.frame}: Returns a dataframe showing the number of unique values in
each column of \code{x}.

\item \code{table}: Coerces \code{x} to \code{data.frame} and passes it to \code{data.frame} method.
}}

\section{Authors}{

\itemize{
\item Desi Quintans (\url{http://www.desiquintans.com})
}
}

\examples{
## --- Default method ---

howmany(rep(letters, 3))
#> 26

howmany(letters, LETTERS)
#> 52

## --- Data.frame method ---

howmany(mtcars)

#> mpg cyl disp hp drat wt qsec vs am gear carb
#>  25   3   27 22   22 29   30  2  2    3    6

howmany(mtcars, -(mpg:disp))

#> hp drat wt qsec vs am gear carb
#> 22   22 29   30  2  2    3    6

howmany(mtcars, drat)

#> drat
#>   22

## --- Table method ---

class(Titanic)
#> [1] "table"

howmany(Titanic)

#> Class Sex Age Survived Freq
#>     4   2   2        2   22

}
