% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_tools.R
\name{if_na}
\alias{if_na}
\title{Return a value or expression if something is NA}
\usage{
if_na(x, yes = TRUE, no = NULL)
}
\arguments{
\item{x}{(Vector) A vector to test.}

\item{yes}{(Any) The value to return if \code{x} is \code{NA}.}

\item{no}{(Any) The value to return if \code{x} is \strong{not} \code{NA}.
Set this to \code{NULL} (its default setting) to return \code{x}.}
}
\value{
The object in \code{yes} or \code{no}, depending on the outcome.
}
\description{
This is just shorthand for \code{ifelse(is.na(x), TRUE, FALSE)} because I like using
that pattern in my \code{dplyr} pipelines.
}
\section{Authors}{

\itemize{
\item Desi Quintans (\url{http://www.desiquintans.com})
}
}

\examples{
vec <- c("hello", NA, "hi")

if_na(vec, "REPLACED")
#> [1] "hello"    "REPLACED" "hi"

if_na(vec, "Was NA", "Was not NA")
#> [1] "Was not NA" "Was NA"     "Was not NA"

}
