% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_tools.R
\name{make_path}
\alias{make_path}
\title{Create a folder path}
\usage{
make_path(...)
}
\arguments{
\item{...}{(Character) Arguments to send to \code{file.path()}. You can provide a complete
path as a single string, or incrementally build a path with many strings.}
}
\value{
(Character) A file path. Automatically adds trailing slashes if required.
}
\description{
Allows you to build a file path using \code{file.path()}, and then makes sure that it
actually exists by creating the folders in the path if needed.
}
\section{Authors}{

\itemize{
\item Desi Quintans (\url{http://www.desiquintans.com})
}
}

\examples{
make_path("path", "subfolder")

#> [1] "path/subfolder"

# And the path/subfolder/ folders were also created in the working directory.

}
