% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_tools.R
\name{mirror_matrix}
\alias{mirror_matrix}
\title{Mirror a matrix horizontally}
\usage{
mirror_matrix(mat, MARGIN = 2)
}
\arguments{
\item{mat}{(Matrix) A matrix object}

\item{MARGIN}{(Integer) `1` mirrors the order of rows, `2` mirrors the order
of columns.}
}
\value{
A version of `mat` with its columns or rows in reversed order. Names are
   preserved if the matrix has any.
}
\description{
Mirror a matrix horizontally
}
\examples{
m <- matrix(1:6, ncol = 2, nrow = 3, byrow = FALSE)
m

#>      [,1] [,2]
#> [1,]    1    4
#> [2,]    2    5
#> [3,]    3    6

# Just like apply(), MARGIN = 1 is rows and MARGIN = 2 is cols.

mirror_matrix(m, 2)

#>      [,1] [,2]
#> [1,]    4    1
#> [2,]    5    2
#> [3,]    6    3

mirror_matrix(m, 1)

#>      [,1] [,2]
#> [1,]    3    6
#> [2,]    2    5
#> [3,]    1    4

}
