% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_tools.R
\name{Mode}
\alias{Mode}
\title{Mode of a vector (numeric/character/factor)}
\usage{
Mode(x, na.rm = FALSE, ties = FALSE, mean = FALSE)
}
\arguments{
\item{x}{(Char/Numeric/Factor) A vector.}

\item{na.rm}{(Logical) If \code{TRUE}, NAs will be silently removed.}

\item{ties}{(Logical) If \code{TRUE} and \code{x} has multiple modes (e.g. \code{c(2, 2, 1, 1)}),
return all of modes that were found (\code{2, 1}). If \code{FALSE}, only the first mode
(the one first appearing in \code{x}) will be returned (\code{2}).}

\item{mean}{(Logical) If \code{TRUE}, return the average of all the modes. Only makes sense
for a numeric vector. Only the first mode is returned if \code{ties == FALSE},}
}
\value{
A vector of the mode value(s).
}
\description{
There is no built-in function to find the mode of something. This function can find
the mode of a numeric, character, or factor vector. It will return multiple values in
the case of a multi-modal dataset. In the case of a numeric vector, it can return a
single value that is the mean of the modes.
}
\details{
If all values are unique, it will return \strong{all} of the values.
}
\section{Authors}{

\itemize{
\item Ken Williams (\url{https://stackoverflow.com/users/169947/ken-williams})
\item jprockbelly (\url{https://stackoverflow.com/users/1502898/jprockbelly})
\item digEmAll (\url{https://stackoverflow.com/users/316644/digemall})
\item Desi Quintans (\url{http://www.desiquintans.com})
}
}

\section{Source}{

\url{https://stackoverflow.com/a/8189441/5578429}
}

\examples{
vec <- c(1, 2, 3, 4, 4, 4, 3, 3, NA, NA, NA)

Mode(vec)
#> [1]  3  4 NA

Mode(vec, na.rm = TRUE)
#> [1] 3 4

Mode(vec, na.rm = FALSE, mean = TRUE)
#> [1] NA

Mode(vec, na.rm = TRUE, mean = TRUE)
#> [1] 3.5

Mode(1:4)
#> [1] 1 2 3 4

Mode(1:4, ties = FALSE)
#> [1] 1

Mode(1:4, mean = TRUE)
#> [1] 2.5

}
