% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{palette_distinct}
\alias{palette_distinct}
\title{A palette of 1,022 visually-distinct colours}
\usage{
palette_distinct(n = NULL, random = FALSE, spaced = FALSE,
  alpha = NULL)
}
\arguments{
\item{n}{(Numeric) The number of colours to deliver.}

\item{random}{(Logical) If \code{FALSE} (default), you will get \code{n} colours in the order
they appear in the list. If \code{TRUE}, you will get \code{n} colours chosen randomly.}

\item{spaced}{(Logical) If \code{TRUE}, \code{n} colours will be chosen equally spaced within the
\code{col_list}, e.g. 3 colours from a list of 16 will return 1, 8, and 16th colours.
If \code{FALSE} (by default), colours will not be spaced out.}

\item{alpha}{(Numeric) A single decimal value that modifies the opacity of the colours.
\code{alpha = 0.65} makes all of the colours 65 percent opaque (35 percent transparent).
If \code{alpha = NULL}, no alpha channel is added and the colours are 100 percent opaque.}
}
\value{
A character vector of hex colours. If \code{alpha = NULL}, the colours will be in
RGB Hex format (e.g. #FFFF00). If \code{alpha} is not \code{NULL}, the colours will be in
RGBA Hex format (e.g. #FFFF00CB).
}
\description{
There's so many colours, man. Many of these are not colourblind-safe! Many also have
poor contrast against each other, but they can still be differentiated as different
colours. Never rely on colour alone to differentiate data points; use shapes, opacity,
patterns, etc. Better yet, consider simplifying your graph so that it doesn't have so
many dimensions.
}
\details{
Code to plot all the colours can be found in the Examples section. You can even use a
colour picker to choose colours from the resulting plot and be assured that they can
be differentiated.
}
\section{Authors}{

\itemize{
\item Tatarize (\url{https://stackoverflow.com/users/631911/tatarize})
\item Desi Quintans (<http://www.desiquintans.com)
}
}

\section{Source}{

\url{https://stackoverflow.com/a/12224359/5578429}
}

\examples{
# To see all of the colours (ordered left-to-right and top-to-bottom):
show_colours(palette_distinct())

# To get the first 4 colours:
palette_distinct(4)

#> [1] "#000000" "#FFFF00" "#1CE6FF" "#FF34FF"

# To pick 4 colours randomly:
palette_distinct(4, random = TRUE)

#> [1] "#2F2E2C" "#DFE3E6" "#5C424D" "#FFE47D"

# To make the colours 75 percent opaque:
palette_distinct(4, random = TRUE, alpha = 0.75)

#> [1] "#4D913EBF" "#2A7FFFBF" "#830055BF" "#664327BF"

}
