% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{palette_distinct}
\alias{palette_distinct}
\title{A palette of 1,022 visually-distinct colours}
\usage{
palette_distinct(n = 1022, random = FALSE)
}
\arguments{
\item{n}{(Numeric) The number of colours to deliver.}

\item{random}{(Logical) If \code{FALSE} (default), you will get \code{n} colours in the order
they appear in the list. If \code{TRUE}, you will get \code{n} colours chosen randomly.}
}
\value{
A character vector of hex colours.
}
\description{
There's so many colours, man. Many of these are not colourblind-safe! Many also have
poor contrast against each other, but they can still be differentiated as different
colours. Never rely on colour alone to differentiate data points; use shapes, opacity,
patterns, etc. Better yet, consider simplifying your graph so that it doesn't have so
many dimensions.
}
\details{
Code to plot all the colours can be found in the Examples section. You can even use a
colour picker to choose colours from the resulting plot and be assured that they can
be differentiated.
}
\section{Authors}{

\itemize{
\item Tatarize (\url{https://stackoverflow.com/users/631911/tatarize})
\item Desi Quintans (<http://www.desiquintans.com)
}
}

\section{Source}{

\url{https://stackoverflow.com/a/12224359/5578429}
}

\examples{
# To see all of the colours (ordered left-to-right and top-to-bottom):
image(apply(matrix(1022:1, ncol = 73, nrow = 14, byrow = TRUE), 1, rev), col = palette_distinct())

# To get the first 4 colours:
palette_distinct(4)

#> [1] "#000000" "#FFFF00" "#1CE6FF" "#FF34FF"

# To pick 4 colours randomly:
palette_distinct(4, random = TRUE)

#> [1] "#2F2E2C" "#DFE3E6" "#5C424D" "#FFE47D"

}
