% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_tools.R
\name{percentile}
\alias{percentile}
\title{Quick percentile overview}
\usage{
percentile(num, cuts = c(0, 0.1, 0.2, 0.25, 0.33, 0.5, 0.66, 0.75, 0.8,
  0.85, 0.9, 0.95, 0.99, 1))
}
\arguments{
\item{num}{(Numeric) A vector.}

\item{cuts}{(Numeric) A vector of percentiles to calculate.}
}
\value{
A named numeric vector of percentiles and their values.
}
\description{
Break down a vector into useful percentiles. If the 25th percentile is 10.5, for
example, then 25 \% of the observations are < 10.5.
}
\section{Authors}{

\itemize{
\item Desi Quintans (\url{http://www.desiquintans.com})
}
}

\examples{
# round(runif(20, min = 0, max = 29))
vec <- c(28, 23, 3, 28, 6, 5, 21, 19, 9, 17, 22, 23, 26, 9, 5, 20, 19, 24, 3, 27)
perc <- percentile(vec)

perc

#>   0\%   10\%   20\%   25\%   33\%   50\%   66\%   75\%   80\%   85\%   90\%   95\%   99\%  100\%
#> 3.00  4.80  5.80  8.25 11.16 19.50 22.54 23.25 24.40 26.15 27.10 28.00 28.00 28.00

perc["66\%"]

#>   66\%
#> 22.54

}
