% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tools.R
\name{show_colours}
\alias{show_colours}
\alias{show_colors}
\title{Preview a list of colours as a grid}
\usage{
show_colours(
  col_list,
  arrange = "rect",
  pad = "#FFFFFF",
  asp = NA,
  main = NULL
)

show_colors(col_list, arrange = "rect", pad = "#FFFFFF", asp = NA, main = NULL)
}
\arguments{
\item{col_list}{(Character) A vector of colours in RGB Hex or RGBA Hex format.}

\item{arrange}{(Character: \code{"rect"}, \code{"rows"}, or \code{"cols"}) By default (\code{"rect"}),
colours are displayed in a rectangular panel. \code{"rows"} arranges them as
horizontal stripes top-to-bottom, and \code{"cols"} arranges them as vertical
stripes left-to-right.}

\item{pad}{(Character) If there are too few colours to fill all of the spaces in
the grid, what colour should be used to pad it out? \code{pad = "last"} repeats the
last colour in the colour list. A hex colour can be provided too. Otherwise,
this defaults to white (`#FFFFFF``).}

\item{asp}{(Numeric or \code{NA}) The aspect ratio of the image. \code{asp = 1} produces a square
plot with square cells. If \code{NA}, \code{plot.window}'s default will be used.}

\item{main}{(Character or \code{NULL}) Leave this as \code{NULL} to generate a title. Supply a
string to define your own title.}
}
\value{
A graphic that shows all of the colours in \code{col_list}.
}
\description{
Preview a list of colours as a grid
}
\examples{
show_colours(colours(distinct = TRUE))

show_colours(colours(distinct = TRUE), arrange = "cols")

}
