% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_programming.R
\name{sift}
\alias{sift}
\title{Fuzzily search column names and labels}
\usage{
sift(df, query, dist = 0.1)
}
\arguments{
\item{df}{(Dataframe) The dataframe}

\item{query}{(Character) A string (or regular expression) to search for.}

\item{dist}{(Numeric) Maximum string distance. Either an integer or a double; see
\code{max.distance} in \code{base::agrep()}. Set to \code{0} to force an exact match.}
}
\value{
Prints out matching name & label combinations, and invisibly returns the same.
}
\description{
When working with dataframes hundreds of columns wide, it can be hard to find which
variable contains a particular measurement. This function lets you fuzzily search a
dataframe's column names and labels (the \code{attr()} named \code{"label"}) for a match. It also
reports appropriate summary statistics like:
}
\details{
\itemize{
\item Variable type
\item Number of unique elements
\item Percentage of the vector that is \code{NA} or \code{NaN}
\item Range (min to max)
\item Mode (most common value(s))
\item Median (50th percentile value)
}
}
\section{Authors}{

\itemize{
\item Desi Quintans (\url{http://www.desiquintans.com})
}
}

\examples{

sift(iris_labelled, "pal.width")
# Note fuzzy searching gets both "pal.width" and "tal.width"

#> Sepal.Width | type: numeric | Uniq: 23 | NA \%: 0
#>     Width of sepals.
#>     range: 2-4.4 | mode: 3 | median: 3
#> Petal.Width | type: numeric | Uniq: 22 | NA \%: 0
#>     Width of petals.
#>     range: 0.1-2.5 | mode: 0.2 | median: 1.3

sift(iris_labelled, "pal.width", dist = 0)
# dist = 0 forces an exact match

#> Sepal.Width | type: numeric | Uniq: 23 | NA \%: 0
#>     Width of sepals.
#>     range: 2-4.4 | mode: 3 | median: 3

sift(iris_labelled, "versicol")
# The columns' labels are also searched.

#> Species | type: factor | Uniq: 3 | NA \%: 0
#>     Species of iris: setosa, versicolor, virginica
#>     mode: setosa, versicolor, and 1 other

}
